refactor(backend): split main.py into routers; hard-fail mounts; single DB module (T8)
Mechanical, verbatim split of backend/main.py: 2,985 lines before, 362 after.
No logic, path, method, status-code, or response changes.

Modules created:
- db.py — DB_URL resolution, the module-level psycopg2 conn (+ None fallback
  and warning), the deed_pdfs ensure block, and get_db_connection(). conn is
  a global that get_db_connection() rebinds on reconnect, so every consumer
  does `import db` and references `db.conn` (module-attribute access);
  `from db import conn` is banned because that snapshot goes stale after a
  reconnect. LAST_REQUEST_TS in app_state.py follows the same rule.
- app_state.py — METRICS defaultdict + LAST_REQUEST_TS shared by the metrics
  middleware (main.py) and /admin/system-metrics (routers/admin_inline.py).
- routers/system.py (2 routes), routers/users_auth.py (9), routers/
  admin_inline.py (11), routers/pricing.py (6), routers/deeds_crud.py (9),
  routers/sharing.py (9), routers/property_cache_routes.py (3) — all
  APIRouter with no prefix; path strings, decorators (including the
  dependencies=[Depends(get_current_admin)] guards) unchanged. Pydantic
  models moved with the module that uses them; unused legacy models and
  helpers (check_plan_limits, compute_doc_transfer_tax, needs_exhibit_a,
  Jinja env, RECORDER_PROFILES) stay in main.py.

Every router mount — old and new — is now a direct import +
app.include_router with no try/except. The old try/except-around-mount
pattern silently dropped routers when an import failed (the T7 incident:
a broken import made routes vanish without any startup error); a broken
router now fails loudly at process start. Mount order preserves the
pre-split registration order (billing first, existing includes in their
original sequence, then the extracted inline-section routers in
file-appearance order), keeping route-matching precedence identical.

Deliberately preserved quirk: the /users/profile handler is named
get_user_profile and shadowed database.get_user_profile in pre-split
main.py; routers/property_cache_routes.py imports the handler from
routers/users_auth.py so /ai/deed-suggestions and /users/profile/enhanced
behave exactly as before.

scripts/six_flow_baseline.py: `from main import app, conn` -> `from main
import app` (conn was imported unused; main no longer defines it).

Verification (T8 gate):
- pytest tests/ (incl. test_openapi_contract.py against the committed
  snapshot, not re-recorded): 28 passed.
- six_flow_baseline.py verify vs local Postgres: "SIX-FLOW BASELINE: all
  flows match the recorded snapshot", exit 0.
- python3 -m py_compile on every created/modified file: OK.
- AST-level diff of every moved function/class against pre-split main.py
  (reversing only conn->db.conn / METRICS->app_state.METRICS /
  @app->@router substitutions): zero differences.
