+++ backend/main.py
@@ -18,12 +18,12 @@ import tempfile
 import json
 from database import (
     create_user, get_user_by_email, create_deed, get_user_deeds,
-    get_user_profile, update_user_profile, get_cached_property, 
+    get_user_profile, update_user_profile, get_cached_property,
     cache_property_data, get_recent_properties
 )
 from ai_assist import suggest_defaults, validate_deed_data
 from auth import (
-    get_password_hash, verify_password, create_access_token, 
+    get_password_hash, verify_password, create_access_token,
     get_current_user_id, get_current_user_email, AuthUtils, get_current_admin
 )
 from middleware.qa_instrumentation import QAInstrumentationMiddleware, get_qa_health_status
@@ -39,156 +39,6 @@ if os.getenv("ENVIRONMENT") == "staging":
         QAInstrumentationMiddleware,
         enable_detailed_logging=True
     )
-
-# PHASE 23-B: Include billing & reporting routers (October 30, 2025)
-try:
-    from phase23_billing.app_include import include_billing_routers
-    include_billing_routers(app)
-    print("✅ Phase 23-B billing & reporting endpoints loaded successfully")
-except ImportError as e:
-    print(f"⚠️ Phase 23-B billing endpoints not available: {e}")
-except Exception as e:
-    print(f"❌ Error loading Phase 23-B billing endpoints: {e}")
-
-# T4: the legacy no-auth ai_router (root ai_assist.py) was removed — its
-# POST /api/ai/assist shadowed api/ai_assist.py's and had no consumers.
-# /api/ai/* is owned solely by api/ai_assist.py (mounted below).
-
-# Include Property Integration API router
-try:
-    from api.property_endpoints import router as property_router
-    app.include_router(property_router)
-    print("✅ Property integration endpoints loaded successfully")
-except ImportError as e:
-    print(f"⚠️ Property integration endpoints not available: {e}")
-    print("⚠️ Core functionality will work without property integration")
-except Exception as e:
-    print(f"❌ Error loading property endpoints: {e}")
-    print("⚠️ Core functionality will work without property integration")
-
-# Include new dynamic wizard API routers
-try:
-    from api.ai_assist import router as ai_assist_router
-    app.include_router(ai_assist_router, prefix="/api/ai", tags=["AI Assistant"])
-    print("✅ AI assist endpoints loaded successfully")
-except ImportError as e:
-    print(f"⚠️ AI assist endpoints not available: {e}")
-except Exception as e:
-    print(f"❌ Error loading AI assist endpoints: {e}")
-
-try:
-    from api.generate_deed import router as generate_deed_router
-    app.include_router(generate_deed_router, prefix="/api", tags=["Document Generation"])
-    print("✅ Document generation endpoints loaded successfully")
-except ImportError as e:
-    print(f"⚠️ Document generation endpoints not available: {e}")
-except Exception as e:
-    print(f"❌ Error loading document generation endpoints: {e}")
-
-# Include new Grant Deed CA router
-try:
-    from routers.deeds import router as deeds_router
-    app.include_router(deeds_router, prefix="/api")
-    print("✅ Grant Deed CA endpoints loaded successfully")
-except ImportError as e:
-    print(f"⚠️ Grant Deed CA endpoints not available: {e}")
-except Exception as e:
-    print(f"❌ Error loading Grant Deed CA endpoints: {e}")
-
-# Include PHASE 8 extra deed types router
-try:
-    from routers.deeds_extra import router as deeds_extra_router
-    ENABLE_EXTRA_DEEDS = os.getenv("ENABLE_DEED_TYPES_EXTRA", "false").lower() == "true"
-    if ENABLE_EXTRA_DEEDS:
-        app.include_router(deeds_extra_router)
-        print("✅ Extra Deed Types endpoints loaded successfully (Quitclaim, Interspousal, Warranty, Tax)")
-    else:
-        print("⚠️ Extra Deed Types endpoints disabled via ENABLE_DEED_TYPES_EXTRA flag")
-except ImportError as e:
-    print(f"⚠️ Extra Deed Types endpoints not available: {e}")
-except Exception as e:
-    print(f"❌ Error loading Extra Deed Types endpoints: {e}")
-
-# Document types API for dynamic wizard
-try:
-    from api.doc_types import router as doc_types_router
-    app.include_router(doc_types_router, prefix="/api", tags=["Doc Types"])
-    print("✅ Document types endpoints loaded successfully")
-except ImportError as e:
-    print(f"⚠️ Document types endpoints not available: {e}")
-except Exception as e:
-    print(f"❌ Error loading document types endpoints: {e}")
-
-# T4: routers/ai.py removed (chain-of-title / profile-request) — no frontend
-# consumers; called TitlePoint's SOAP endpoint directly, bypassing the service.
-
-# AUTH HARDENING: Password reset, email verification, refresh tokens
-try:
-    from routers.auth_extra import router as auth_extra_router
-    app.include_router(auth_extra_router, prefix="", tags=["Auth Extra"])
-    print("✅ Auth hardening endpoints loaded successfully (password reset, email verification)")
-except ImportError as e:
-    print(f"⚠️ Auth hardening endpoints not available: {e}")
-except Exception as e:
-    print(f"❌ Error loading auth hardening endpoints: {e}")
-
-# ADMIN HONESTY PASS: Real admin endpoints with pagination, search, exports
-try:
-    from routers.admin_api_v2 import router as admin_api_v2_router
-    app.include_router(admin_api_v2_router, prefix="", tags=["Admin v2"])
-    print("✅ Admin v2 endpoints loaded successfully (users/deeds search, detail, exports)")
-except ImportError as e:
-    print(f"⚠️ Admin v2 endpoints not available: {e}")
-except Exception as e:
-    print(f"❌ Error loading Admin v2 endpoints: {e}")
-
-# PHASE 7.5: Notifications System (gap-plan)
-try:
-    from routers.notifications import router as notifications_router
-    app.include_router(notifications_router, prefix="/notifications", tags=["Notifications"])
-    print("✅ Phase 7.5: Notifications system loaded (feature flag: NOTIFICATIONS_ENABLED)")
-except ImportError as e:
-    print(f"⚠️ Phase 7.5: Notifications not available: {e}")
-except Exception as e:
-    print(f"❌ Phase 7.5: Error loading Notifications: {e}")
-
-# T3: the flag-gated "enhanced sharing" and "/deed-shares feedback" routers
-# were removed — no frontend consumer ever called them. The live sharing
-# stack is the /shared-deeds + /approve/{token} endpoints below (deed_shares
-# table), which now also serve feedback and revoke.
-
-# PATCH 5: Industry Partners API (Full Stack - org-scoped with admin)
-try:
-    from routers.partners import router as partners_router
-    from routers.admin_partners import router as admin_partners_router
-    app.include_router(partners_router, prefix="/partners", tags=["Partners"])
-    app.include_router(admin_partners_router, prefix="/admin/partners", tags=["Admin Partners"])
-    print("✅ Patch 5: Industry Partners API loaded (org-scoped, user CRUD + admin panel)")
-except ImportError as e:
-    print(f"⚠️ Patch 5: Partners API not available: {e}")
-except Exception as e:
-    print(f"❌ Phase 15 v5: Error loading Industry Partners API: {e}")
-
-# QR Verification System (Public endpoint - no auth required)
-try:
-    from routers.verification import router as verification_router, admin_router as verification_admin_router
-    app.include_router(verification_router, tags=["Document Verification"])
-    app.include_router(verification_admin_router, tags=["Admin Verification"])
-    print("✅ QR Verification System loaded (/api/verify/{code} - public, /admin/verification/* - admin)")
-except ImportError as e:
-    print(f"⚠️ QR Verification System not available: {e}")
-except Exception as e:
-    print(f"❌ QR Verification System error: {e}")
-
-# Public API v1 (Partner API for deed generation)
-try:
-    from routers.api_v1.router import router as api_v1_router
-    app.include_router(api_v1_router, tags=["Public API v1"])
-    print("✅ Public API v1 loaded (/api/v1/deeds - partner deed generation)")
-except ImportError as e:
-    print(f"⚠️ Public API v1 not available: {e}")
-except Exception as e:
-    print(f"❌ Public API v1 error: {e}")
 
 # Allow CORS for local dev and frontend
 app.add_middleware(
@@ -208,70 +58,145 @@ app.add_middleware(
 @app.middleware("http")
 async def metrics_middleware(request: Request, call_next):
     """Track request metrics for admin monitoring"""
-    global LAST_REQUEST_TS
     t0 = time()
-    METRICS['requests_total'] += 1
+    app_state.METRICS['requests_total'] += 1
     try:
         response = await call_next(request)
-        METRICS[f"status_{response.status_code}"] += 1
+        app_state.METRICS[f"status_{response.status_code}"] += 1
         return response
     finally:
         dur = time() - t0
-        METRICS['latency_ms_sum'] += int(dur * 1000)
-        LAST_REQUEST_TS = time()
+        app_state.METRICS['latency_ms_sum'] += int(dur * 1000)
+        app_state.LAST_REQUEST_TS = time()
 
 # Stripe configuration
 stripe.api_key = os.getenv("STRIPE_SECRET_KEY")
 
-# Database connection with auto-reconnect support
-DB_URL = os.getenv("DATABASE_URL") or os.getenv("DB_URL")
-if DB_URL:
-    conn = psycopg2.connect(DB_URL)
+# T8: DB connection (module-level conn + get_db_connection) lives in db.py.
+# Always access the connection as db.conn — get_db_connection() rebinds the
+# global on reconnect, so a from-import would go stale.
+import db
+
+# T8: shared mutable metrics state lives in app_state.py (module-attribute
+# access required — LAST_REQUEST_TS is rebound on every request).
+import app_state
+
+# ============================================================================
+# Router mounts (T8): every mount is a DIRECT import + include_router.
+# The old try/except-around-mount pattern silently skipped routers when an
+# import failed (the T7 incident) — a broken router must now fail loudly at
+# startup instead of quietly dropping its routes.
+# Mount order preserves the pre-split registration order: billing first,
+# then the existing includes in their original sequence, then the routers
+# extracted from main.py's inline sections in the order those sections
+# appeared (decorator routes were registered after all includes).
+# ============================================================================
+
+# PHASE 23-B: Include billing & reporting routers (October 30, 2025)
+from phase23_billing.app_include import include_billing_routers
+include_billing_routers(app)
+print("✅ Phase 23-B billing & reporting endpoints loaded successfully")
+
+# T4: the legacy no-auth ai_router (root ai_assist.py) was removed — its
+# POST /api/ai/assist shadowed api/ai_assist.py's and had no consumers.
+# /api/ai/* is owned solely by api/ai_assist.py (mounted below).
+
+# Include Property Integration API router
+from api.property_endpoints import router as property_router
+app.include_router(property_router)
+print("✅ Property integration endpoints loaded successfully")
+
+# Include new dynamic wizard API routers
+from api.ai_assist import router as ai_assist_router
+app.include_router(ai_assist_router, prefix="/api/ai", tags=["AI Assistant"])
+print("✅ AI assist endpoints loaded successfully")
+
+from api.generate_deed import router as generate_deed_router
+app.include_router(generate_deed_router, prefix="/api", tags=["Document Generation"])
+print("✅ Document generation endpoints loaded successfully")
+
+# Include new Grant Deed CA router
+from routers.deeds import router as deeds_router
+app.include_router(deeds_router, prefix="/api")
+print("✅ Grant Deed CA endpoints loaded successfully")
+
+# Include PHASE 8 extra deed types router
+from routers.deeds_extra import router as deeds_extra_router
+ENABLE_EXTRA_DEEDS = os.getenv("ENABLE_DEED_TYPES_EXTRA", "false").lower() == "true"
+if ENABLE_EXTRA_DEEDS:
+    app.include_router(deeds_extra_router)
+    print("✅ Extra Deed Types endpoints loaded successfully (Quitclaim, Interspousal, Warranty, Tax)")
 else:
-    conn = None
-    print("Warning: No database connection URL found")
+    print("⚠️ Extra Deed Types endpoints disabled via ENABLE_DEED_TYPES_EXTRA flag")
 
-# T2: make sure the stored-PDF table exists (idempotent; no Alembic in this repo)
-if conn:
-    try:
-        from services.deed_pdf import ensure_deed_pdfs_table
-        ensure_deed_pdfs_table(conn)
-    except Exception as _pdf_table_error:
-        print(f"Warning: could not ensure deed_pdfs table: {_pdf_table_error}")
+# Document types API for dynamic wizard
+from api.doc_types import router as doc_types_router
+app.include_router(doc_types_router, prefix="/api", tags=["Doc Types"])
+print("✅ Document types endpoints loaded successfully")
 
-def get_db_connection():
-    """Get database connection, reconnecting if necessary"""
-    global conn
-    if not DB_URL:
-        raise HTTPException(status_code=500, detail="Database connection not available")
-    
-    try:
-        # Test if connection is alive
-        if conn is None or conn.closed:
-            print("⚠️ Database connection closed, reconnecting...")
-            conn = psycopg2.connect(DB_URL)
-            print("✅ Database reconnected successfully")
-        else:
-            # Test with a simple query
-            with conn.cursor() as cur:
-                cur.execute("SELECT 1")
-    except (psycopg2.OperationalError, psycopg2.InterfaceError, AttributeError) as e:
-        print(f"⚠️ Database connection lost ({e}), reconnecting...")
-        try:
-            conn = psycopg2.connect(DB_URL)
-            print("✅ Database reconnected successfully")
-        except Exception as reconnect_error:
-            print(f"❌ Failed to reconnect to database: {reconnect_error}")
-            raise HTTPException(status_code=500, detail="Database connection failed")
-    
-    return conn
+# T4: routers/ai.py removed (chain-of-title / profile-request) — no frontend
+# consumers; called TitlePoint's SOAP endpoint directly, bypassing the service.
 
-# Phase 6-2: System Metrics Tracking
-METRICS = defaultdict(int)
-LAST_REQUEST_TS = 0.0
+# AUTH HARDENING: Password reset, email verification, refresh tokens
+from routers.auth_extra import router as auth_extra_router
+app.include_router(auth_extra_router, prefix="", tags=["Auth Extra"])
+print("✅ Auth hardening endpoints loaded successfully (password reset, email verification)")
 
-# Phase 6-2: Draft persistence (in-memory, replace with DB in Phase 6-3)
-_DRAFTS = {}
+# ADMIN HONESTY PASS: Real admin endpoints with pagination, search, exports
+from routers.admin_api_v2 import router as admin_api_v2_router
+app.include_router(admin_api_v2_router, prefix="", tags=["Admin v2"])
+print("✅ Admin v2 endpoints loaded successfully (users/deeds search, detail, exports)")
+
+# PHASE 7.5: Notifications System (gap-plan)
+from routers.notifications import router as notifications_router
+app.include_router(notifications_router, prefix="/notifications", tags=["Notifications"])
+print("✅ Phase 7.5: Notifications system loaded (feature flag: NOTIFICATIONS_ENABLED)")
+
+# T3: the flag-gated "enhanced sharing" and "/deed-shares feedback" routers
+# were removed — no frontend consumer ever called them. The live sharing
+# stack is the /shared-deeds + /approve/{token} endpoints (routers/sharing.py),
+# (deed_shares table), which now also serve feedback and revoke.
+
+# PATCH 5: Industry Partners API (Full Stack - org-scoped with admin)
+from routers.partners import router as partners_router
+from routers.admin_partners import router as admin_partners_router
+app.include_router(partners_router, prefix="/partners", tags=["Partners"])
+app.include_router(admin_partners_router, prefix="/admin/partners", tags=["Admin Partners"])
+print("✅ Patch 5: Industry Partners API loaded (org-scoped, user CRUD + admin panel)")
+
+# QR Verification System (Public endpoint - no auth required)
+from routers.verification import router as verification_router, admin_router as verification_admin_router
+app.include_router(verification_router, tags=["Document Verification"])
+app.include_router(verification_admin_router, tags=["Admin Verification"])
+print("✅ QR Verification System loaded (/api/verify/{code} - public, /admin/verification/* - admin)")
+
+# Public API v1 (Partner API for deed generation)
+from routers.api_v1.router import router as api_v1_router
+app.include_router(api_v1_router, tags=["Public API v1"])
+print("✅ Public API v1 loaded (/api/v1/deeds - partner deed generation)")
+
+# T8: routers extracted from main.py's former inline sections, mounted in the
+# order the sections appeared in the pre-split file.
+from routers.system import router as system_router
+app.include_router(system_router)
+
+from routers.users_auth import router as users_auth_router
+app.include_router(users_auth_router)
+
+from routers.admin_inline import router as admin_inline_router
+app.include_router(admin_inline_router)
+
+from routers.deeds_crud import router as deeds_crud_router
+app.include_router(deeds_crud_router)
+
+from routers.sharing import router as sharing_router
+app.include_router(sharing_router)
+
+from routers.pricing import router as pricing_router
+app.include_router(pricing_router)
+
+from routers.property_cache_routes import router as property_cache_routes_router
+app.include_router(property_cache_routes_router)
 
 # Jinja2 environment for deed templates
 # Use absolute path that works both locally and on Render
@@ -355,39 +280,6 @@ class UserUpdate(BaseModel):
     city: Optional[str] = None
     country: Optional[str] = None
 
-class AdminUserUpdate(BaseModel):
-    first_name: Optional[str] = None
-    last_name: Optional[str] = None
-    username: Optional[str] = None
-    city: Optional[str] = None
-    country: Optional[str] = None
-    is_active: Optional[bool] = None
-    subscription_status: Optional[str] = None
-
-class DeedCreate(BaseModel):
-    deed_type: str = Field(..., description="Deed type, e.g., 'grant-deed'")
-    property_address: Optional[str] = Field(default=None)
-    apn: Optional[str] = Field(default=None)
-    county: Optional[str] = Field(default=None)
-    legal_description: str = Field(..., min_length=1, description="Legal description (required, non-empty)")
-    owner_type: Optional[str] = Field(default=None)
-    sales_price: Optional[float] = Field(default=None)
-    grantor_name: str = Field(..., min_length=1, description="Grantor name (required, non-empty)")
-    grantee_name: str = Field(..., min_length=1, description="Grantee name (required, non-empty)")
-    vesting: Optional[str] = Field(default=None)
-    requested_by: Optional[str] = Field(default=None, description="Person/company requesting the deed (e.g., escrow officer)")
-    source: Optional[str] = Field(default=None, description="Data source tracking (e.g., 'modern-canonical', 'classic')")
-    # T2: extras persisted into deeds.metadata so the stored PDF can render
-    # the complete document (DTT declaration, reference numbers, mail-to).
-    dtt: Optional[Dict] = Field(default=None, description="Documentary transfer tax details from the builder")
-    title_order_no: Optional[str] = Field(default=None)
-    escrow_no: Optional[str] = Field(default=None)
-    return_to: Optional[str] = Field(default=None, description="Mail-to name for the recorded deed")
-    provenance: Optional[Dict] = Field(default=None, description="Per-field source + confirmation timestamps (Ticket B)")
-
-    class Config:
-        extra = "ignore"  # Ignore extra fields from frontend
-
 class PaymentMethodCreate(BaseModel):
     payment_method_id: str
     set_as_default: bool = False
@@ -401,54 +293,9 @@ class RecipientsCreate(BaseModel):
     deed_id: int
     recipients: List[Recipient]
 
-class ShareDeedCreate(BaseModel):
-    deed_id: int
-    recipient_name: str
-    recipient_email: str
-    recipient_role: str
-    message: Optional[str] = None
-    expires_in_hours: Optional[int] = 168  # Default 7 days (168 hours)
-    
-    @field_validator('expires_in_hours')
-    @classmethod
-    def validate_expiry(cls, v):
-        if v is not None and (v < 1 or v > 720):  # 1 hour to 30 days
-            raise ValueError('Expiration must be between 1 and 720 hours')
-        return v
-
-class ApprovalResponse(BaseModel):
-    approved: bool
-    comments: Optional[str] = None
-
-# New models for registration and plan management
-class UserRegister(BaseModel):
-    email: str
-    password: str
-    confirm_password: str
-    full_name: str
-    role: str
-    company_name: Optional[str] = None
-    company_type: Optional[str] = None
-    phone: Optional[str] = None
-    state: str
-    agree_terms: bool
-    subscribe: bool = False
-
 class DeedData(BaseModel):
     deed_type: str
     data: dict
-
-# Phase 6-2: Draft persistence model
-class DraftPayload(BaseModel):
-    deed_type: str
-    data: dict
-
-class UserLogin(BaseModel):
-    email: str
-    password: str
-
-class UpgradeRequest(BaseModel):
-    plan: str  # 'professional', 'enterprise'
 
 class PlanLimits(BaseModel):
     max_deeds_per_month: int
@@ -457,2529 +304,59 @@ class PlanLimits(BaseModel):
     integrations_enabled: bool
     priority_support: bool
 
-# Health check
-@app.get("/health")
-def health():
-    return {"status": "ok", "message": "DeedPro API is running"}
-
-@app.get("/health/qa")
-async def qa_health_check():
-    """QA-specific health check with instrumentation metrics"""
-    if os.getenv("ENVIRONMENT") != "staging":
-        raise HTTPException(status_code=404, detail="QA health endpoint only available in staging")
-    
-    return get_qa_health_status()
-
-# ============================================================================
-# AUTHENTICATION & REGISTRATION ENDPOINTS
-# ============================================================================
-
-@app.post("/users/register")
-async def register_user(user: UserRegister = Body(...)):
-    """Register a new user with comprehensive validation"""
-    try:
-        # Validation
-        if not user.agree_terms:
-            raise HTTPException(status_code=400, detail="Must agree to terms and conditions")
-        
-        if user.password != user.confirm_password:
-            raise HTTPException(status_code=400, detail="Passwords do not match")
-        
-        # Validate password strength
-        is_valid, message = AuthUtils.validate_password_strength(user.password)
-        if not is_valid:
-            raise HTTPException(status_code=400, detail=message)
-        
-        # Validate email format
-        if not AuthUtils.validate_email(user.email):
-            raise HTTPException(status_code=400, detail="Invalid email format")
-        
-        # Validate state code
-        if not AuthUtils.validate_state_code(user.state):
-            raise HTTPException(status_code=400, detail="Invalid state code")
-        
-        # Hash password
-        hashed_password = get_password_hash(user.password)
-        
-        # PHASE 24-G FIX: Use resilient connection that auto-reconnects
-        conn = get_db_connection()
-        
-        # Insert user into database
-        new_user_id = None
-        with conn.cursor() as cur:
-            cur.execute("""
-                INSERT INTO users (email, password_hash, full_name, role, company_name, 
-                                 company_type, phone, state, subscribe, plan)
-                VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
-                RETURNING id
-            """, (
-                user.email.lower(), hashed_password, user.full_name, user.role,
-                user.company_name, user.company_type, user.phone, user.state.upper(),
-                user.subscribe, 'free'
-            ))
-            result = cur.fetchone()
-            if result:
-                new_user_id = result[0]
-            conn.commit()
-        
-        # Phase 7: Send admin notification for new user registration
-        try:
-            from utils.notifications import notify_new_user_registration
-            
-            admin_email = os.getenv('ADMIN_EMAIL', 'admin@deedpro.com')
-            
-            notification_sent = notify_new_user_registration(
-                admin_email=admin_email,
-                user_email=user.email.lower(),
-                user_name=user.full_name,
-                user_id=new_user_id or 0
-            )
-            
-            if notification_sent:
-                print(f"[Phase 7] ✅ Admin notification sent for new user: {user.email}")
-            else:
-                print(f"[Phase 7] ⚠️ Failed to send admin notification for new user")
-        except Exception as notif_error:
-            # Don't fail registration if notification fails
-            print(f"[Phase 7] ⚠️ Admin notification error (non-blocking): {notif_error}")
-        
-        return {
-            "message": "User registered successfully", 
-            "email": user.email,
-            "plan": "free"
-        }
-        
-    except psycopg2.IntegrityError as ie:
-        # PHASE 24-G FIX: Safe rollback that handles closed connections
-        try:
-            if conn and not conn.closed:
-                conn.rollback()
-        except Exception as rollback_error:
-            print(f"[REGISTER ROLLBACK ERROR] {rollback_error}")
-        raise HTTPException(status_code=400, detail="Email already exists")
-    except Exception as e:
-        # PHASE 24-G FIX: Safe rollback that handles closed connections
-        try:
-            if conn and not conn.closed:
-                conn.rollback()
-        except Exception as rollback_error:
-            print(f"[REGISTER ROLLBACK ERROR] {rollback_error}")
-        raise HTTPException(status_code=500, detail=f"Registration failed: {str(e)}")
-
-@app.post("/users/login")
-async def login_user(credentials: UserLogin = Body(...)):
-    """Authenticate user and return JWT token"""
-    try:
-        # PHASE 24-G FIX: Use resilient connection that auto-reconnects
-        conn = get_db_connection()
-        
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT id, password_hash, full_name, plan, is_active, role 
-                FROM users WHERE email = %s
-            """, (credentials.email.lower(),))
-            user = cur.fetchone()
-        
-        if not user:
-            raise HTTPException(status_code=401, detail="Invalid email or password")
-        
-        # Phase 7.5 FIX: Handle both RealDictCursor (dict) and regular cursor (tuple)
-        if isinstance(user, dict):
-            user_id = user.get('id')
-            password_hash = user.get('password_hash')
-            full_name = user.get('full_name')
-            plan = user.get('plan')
-            is_active = user.get('is_active')
-            role = user.get('role')
-        else:
-            user_id, password_hash, full_name, plan, is_active, role = user
-        
-        if not is_active:
-            raise HTTPException(status_code=401, detail="Account is deactivated")
-        
-        if not verify_password(credentials.password, password_hash):
-            raise HTTPException(status_code=401, detail="Invalid email or password")
-        
-        # Update last login
-        with conn.cursor() as cur:
-            cur.execute("UPDATE users SET last_login = CURRENT_TIMESTAMP WHERE id = %s", (user_id,))
-            conn.commit()
-        
-        # Create access token (AdminFix: include role for admin access)
-        access_token = create_access_token(
-            data={
-                "sub": str(user_id), 
-                "email": credentials.email.lower(),
-                "role": role or "user"  # Default to "user" if role is None
-            }
-        )
-        
-        return {
-            "access_token": access_token,
-            "token_type": "bearer",
-            "user": {
-                "id": user_id,
-                "email": credentials.email.lower(),
-                "full_name": full_name,
-                "plan": plan
-            }
-        }
-        
-    except HTTPException:
-        raise
-    except Exception as e:
-        # PHASE 24-G FIX: Safe rollback that handles closed connections
-        try:
-            if conn and not conn.closed:
-                conn.rollback()
-        except Exception as rollback_error:
-            print(f"[LOGIN ROLLBACK ERROR] {rollback_error}")
-        
-        print(f"[LOGIN ERROR] {e}")
-        raise HTTPException(status_code=500, detail=f"Login failed: {str(e)}")
-
-@app.get("/users/profile")
-async def get_user_profile(user_id: int = Depends(get_current_user_id)):
-    """Get current user's profile information"""
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT id, email, full_name, role, company_name, company_type, 
-                       phone, state, plan, created_at, last_login
-                FROM users WHERE id = %s AND is_active = TRUE
-            """, (user_id,))
-            user = cur.fetchone()
-        
-        if not user:
-            raise HTTPException(status_code=404, detail="User not found")
-        
-        # Get plan limits
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT max_deeds_per_month, api_calls_per_month, ai_assistance, 
-                       integrations_enabled, priority_support
-                FROM plan_limits WHERE plan_name = %s
-            """, (user[8],))  # user[8] is plan
-            limits = cur.fetchone()
-        
-        return {
-            "id": user[0],
-            "email": user[1],
-            "full_name": user[2],
-            "role": user[3],
-            "company_name": user[4],
-            "company_type": user[5],
-            "phone": user[6],
-            "state": user[7],
-            "plan": user[8],
-            "created_at": user[9],
-            "last_login": user[10],
-            "plan_limits": {
-                "max_deeds_per_month": limits[0] if limits else 5,
-                "api_calls_per_month": limits[1] if limits else 100,
-                "ai_assistance": limits[2] if limits else True,
-                "integrations_enabled": limits[3] if limits else False,
-                "priority_support": limits[4] if limits else False
-            }
-        }
-        
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        raise HTTPException(status_code=500, detail=f"Failed to get profile: {str(e)}")
-
-@app.post("/users/upgrade")
-async def upgrade_plan(req: UpgradeRequest, user_id: int = Depends(get_current_user_id)):
-    """Initiate plan upgrade via Stripe Checkout"""
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        # Get user info
-        with conn.cursor() as cur:
-            cur.execute("SELECT stripe_customer_id, email, full_name FROM users WHERE id = %s", (user_id,))
-            user = cur.fetchone()
-        
-        if not user:
-            raise HTTPException(status_code=404, detail="User not found")
-        
-        customer_id, email, full_name = user
-        
-        # Create Stripe customer if not exists
-        if not customer_id:
-            customer = stripe.Customer.create(
-                email=email,
-                name=full_name,
-                metadata={"user_id": str(user_id)}
-            )
-            customer_id = customer.id
-            
-            # Update database with customer ID
-            with conn.cursor() as cur:
-                cur.execute("UPDATE users SET stripe_customer_id = %s WHERE id = %s", (customer_id, user_id))
-                conn.commit()
-        
-        # Map plans to Stripe price IDs (create these in your Stripe dashboard)
-        price_map = {
-            'professional': os.getenv('STRIPE_PROFESSIONAL_PRICE_ID', 'price_professional_default'),
-            'enterprise': os.getenv('STRIPE_ENTERPRISE_PRICE_ID', 'price_enterprise_default')
-        }
-        
-        if req.plan not in price_map:
-            raise HTTPException(status_code=400, detail="Invalid plan selection")
-        
-        # Create Stripe Checkout session
-        session = stripe.checkout.Session.create(
-            customer=customer_id,
-            client_reference_id=str(user_id),
-            metadata={'plan': req.plan, 'user_id': str(user_id)},
-            payment_method_types=['card'],
-            line_items=[{
-                'price': price_map[req.plan],
-                'quantity': 1
-            }],
-            mode='subscription',
-            success_url=f"{os.getenv('FRONTEND_URL', 'http://localhost:3000')}/account-settings?success=true",
-            cancel_url=f"{os.getenv('FRONTEND_URL', 'http://localhost:3000')}/account-settings?canceled=true",
-            allow_promotion_codes=True,
-        )
-        
-        return {"session_url": session.url, "session_id": session.id}
-        
-    except stripe.error.StripeError as e:
-        raise HTTPException(status_code=400, detail=f"Stripe error: {str(e)}")
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Upgrade failed: {str(e)}")
-
-# ============================================================================
-# STRIPE WEBHOOK & PAYMENT ENDPOINTS
-# ============================================================================
-
-# NOTE: POST /payments/webhook is handled by phase23_billing/router_webhook.py
-# (registered first in include_billing_routers). The legacy inline handler that
-# lived here was shadowed/dead; its users.plan sync logic has been ported there.
-
-@app.post("/payments/create-portal-session")
-async def create_portal_session(user_id: int = Depends(get_current_user_id)):
-    """Create Stripe customer portal session for subscription management"""
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            cur.execute("SELECT stripe_customer_id FROM users WHERE id = %s", (user_id,))
-            result = cur.fetchone()
-        
-        if not result or not result[0]:
-            raise HTTPException(status_code=404, detail="No billing information found")
-        
-        session = stripe.billing_portal.Session.create(
-            customer=result[0],
-            return_url=f"{os.getenv('FRONTEND_URL', 'http://localhost:3000')}/account-settings"
-        )
-        
-        return {"url": session.url}
-        
-    except stripe.error.StripeError as e:
-        raise HTTPException(status_code=400, detail=f"Stripe error: {str(e)}")
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Portal session creation failed: {str(e)}")
-
 # ============================================================================
 # PLAN LIMITS & USAGE TRACKING
 # ============================================================================
 
 def check_plan_limits(user_id: int, action: str = "deed_creation") -> dict:
     """Check if user has reached plan limits"""
-    if not conn:
+    if not db.conn:
         return {"allowed": True, "message": "Database not available"}
-    
+
     try:
-        with conn.cursor() as cur:
+        with db.conn.cursor() as cur:
             # Get user plan
             cur.execute("SELECT plan FROM users WHERE id = %s", (user_id,))
             result = cur.fetchone()
             if not result:
                 return {"allowed": False, "message": "User not found"}
-            
+
             plan = result[0]
-            
+
             # Get plan limits
             cur.execute("""
-                SELECT max_deeds_per_month, api_calls_per_month 
+                SELECT max_deeds_per_month, api_calls_per_month
                 FROM plan_limits WHERE plan_name = %s
             """, (plan,))
             limits = cur.fetchone()
-            
+
             if not limits:
                 return {"allowed": True, "message": "No limits configured"}
-            
+
             max_deeds, max_api_calls = limits
-            
+
             if action == "deed_creation" and max_deeds > 0:
                 # Check monthly deed count
                 cur.execute("""
-                    SELECT COUNT(*) FROM deeds 
+                    SELECT COUNT(*) FROM deeds
                     WHERE user_id = %s AND created_at >= DATE_TRUNC('month', CURRENT_DATE)
                 """, (user_id,))
                 deed_count = cur.fetchone()[0]
-                
+
                 if deed_count >= max_deeds:
                     return {
-                        "allowed": False, 
+                        "allowed": False,
                         "message": f"Monthly deed limit reached ({max_deeds}). Upgrade your plan for unlimited deeds.",
                         "current_usage": deed_count,
                         "limit": max_deeds
                     }
-            
+
             return {"allowed": True, "message": "Within limits"}
-            
+
     except Exception as e:
         print(f"Limit check error: {str(e)}")
         return {"allowed": True, "message": "Limit check failed, allowing action"}
 
-# ============================================================================
-# ADMIN ENDPOINTS - Platform Management
-# ============================================================================
-
-@app.get("/admin/dashboard", dependencies=[Depends(get_current_admin)])
-def admin_dashboard():
-    """Get admin dashboard overview with key metrics"""
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            # Get total users
-            cur.execute("SELECT COUNT(*) FROM users")
-            total_users = cur.fetchone()[0]
-            
-            # Get active users (logged in within 30 days)
-            cur.execute("""
-                SELECT COUNT(*) FROM users 
-                WHERE is_active = TRUE AND (last_login > NOW() - INTERVAL '30 days' OR last_login IS NULL)
-            """)
-            active_users = cur.fetchone()[0]
-            
-            # Get total deeds
-            cur.execute("SELECT COUNT(*) FROM deeds")
-            total_deeds = cur.fetchone()[0]
-            
-            # Get deeds this month
-            cur.execute("""
-                SELECT COUNT(*) FROM deeds 
-                WHERE created_at >= DATE_TRUNC('month', CURRENT_DATE)
-            """)
-            deeds_this_month = cur.fetchone()[0]
-            
-            # Get subscription breakdown
-            cur.execute("""
-                SELECT plan, COUNT(*) 
-                FROM users 
-                WHERE is_active = TRUE 
-                GROUP BY plan
-            """)
-            plan_counts = dict(cur.fetchall())
-            
-            # Calculate estimated revenue (simplified calculation)
-            professional_users = plan_counts.get('professional', 0)
-            enterprise_users = plan_counts.get('enterprise', 0)
-            monthly_revenue = (professional_users * 29.99) + (enterprise_users * 99.99)
-            total_revenue = monthly_revenue * 12  # Simplified annual estimate
-            
-            # Get recent activity (recent user signups and deed creations)
-            cur.execute("""
-                SELECT 'user_signup' as type, email as user, created_at as timestamp, NULL as deed_id
-                FROM users 
-                WHERE created_at >= NOW() - INTERVAL '7 days'
-                UNION ALL
-                SELECT 'deed_created' as type, u.email as user, d.created_at as timestamp, d.id as deed_id
-                FROM deeds d
-                JOIN users u ON d.user_id = u.id
-                WHERE d.created_at >= NOW() - INTERVAL '7 days'
-                ORDER BY timestamp DESC
-                LIMIT 10
-            """)
-            recent_activity = []
-            for row in cur.fetchall():
-                activity = {
-                    "type": row[0],
-                    "user": row[1],
-                    "timestamp": row[2].isoformat() if row[2] else None
-                }
-                if row[3]:  # deed_id
-                    activity["deed_id"] = row[3]
-                recent_activity.append(activity)
-        
-        dashboard_data = {
-            "total_users": total_users,
-            "active_users": active_users,
-            "total_deeds": total_deeds,
-            "deeds_this_month": deeds_this_month,
-            "total_revenue": total_revenue,
-            "monthly_revenue": monthly_revenue,
-            "subscription_breakdown": {
-                "free": plan_counts.get('free', 0),
-                "professional": plan_counts.get('professional', 0),
-                "enterprise": plan_counts.get('enterprise', 0)
-            },
-            "recent_activity": recent_activity,
-            "growth_metrics": {
-                "user_growth_rate": 0.0,  # Would need historical data
-                "revenue_growth_rate": 0.0,  # Would need historical data
-                "deed_completion_rate": 100.0 if total_deeds > 0 else 0  # Simplified
-            }
-        }
-        
-        return dashboard_data
-        
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching admin dashboard: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch dashboard data: {str(e)}")
-
-@app.get("/admin/users", dependencies=[Depends(get_current_admin)])
-def admin_list_all_users(
-    page: int = Query(1, ge=1),
-    limit: int = Query(50, ge=1, le=100),
-    search: Optional[str] = None,
-    status: Optional[str] = None
-):
-    """List all users with pagination and filtering"""
-    # Query real user data from database
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            # Build query with filters
-            where_conditions = ["TRUE"]
-            params = []
-            
-            if search:
-                where_conditions.append("(email ILIKE %s OR full_name ILIKE %s)")
-                params.extend([f"%{search}%", f"%{search}%"])
-            
-            if status == "active":
-                where_conditions.append("is_active = TRUE")
-            elif status == "inactive":
-                where_conditions.append("is_active = FALSE")
-            
-            where_clause = " AND ".join(where_conditions)
-            
-            # Get total count for pagination
-            count_query = f"SELECT COUNT(*) FROM users WHERE {where_clause}"
-            cur.execute(count_query, params)
-            total_count = cur.fetchone()[0]
-            
-            # Get users with pagination
-            offset = (page - 1) * limit
-            query = f"""
-                SELECT id, email, full_name, role, company_name, phone, state, plan, 
-                       created_at, last_login, is_active
-                FROM users 
-                WHERE {where_clause}
-                ORDER BY created_at DESC
-                LIMIT %s OFFSET %s
-            """
-            params.extend([limit, offset])
-            cur.execute(query, params)
-            users = cur.fetchall()
-            
-            # Get deed counts for each user
-            user_ids = [user[0] for user in users]
-            deed_counts = {}
-            if user_ids:
-                cur.execute("""
-                    SELECT user_id, COUNT(*) 
-                    FROM deeds 
-                    WHERE user_id = ANY(%s)
-                    GROUP BY user_id
-                """, (user_ids,))
-                deed_counts = dict(cur.fetchall())
-            
-            # Format users for response
-            formatted_users = []
-            for user in users:
-                # Split full_name into first and last name
-                full_name = user[2] or ""
-                name_parts = full_name.split(" ", 1)
-                first_name = name_parts[0] if name_parts else ""
-                last_name = name_parts[1] if len(name_parts) > 1 else ""
-                
-                formatted_users.append({
-                    "id": user[0],
-                    "email": user[1],
-                    "first_name": first_name,
-                    "last_name": last_name,
-                    "full_name": user[2],
-                    "role": user[3],
-                    "company_name": user[4],
-                    "phone": user[5],
-                    "state": user[6],
-                    "subscription_plan": user[7],
-                    "subscription_status": "active" if user[10] else "inactive",
-                    "created_at": user[8].isoformat() if user[8] else None,
-                    "last_login": user[9].isoformat() if user[9] else None,
-                    "is_active": user[10],
-                    "total_deeds": deed_counts.get(user[0], 0),
-                    "monthly_revenue": 29.99 if user[7] == "professional" else 99.99 if user[7] == "enterprise" else 0.00
-                })
-            
-            return {
-                "users": formatted_users,
-                "total": total_count,
-                "page": page,
-                "limit": limit,
-                "total_pages": (total_count + limit - 1) // limit
-            }
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching admin users: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch users: {str(e)}")
-
-@app.get("/admin/users/{user_id}", dependencies=[Depends(get_current_admin)])
-def admin_get_user_details(user_id: int):
-    """Get detailed information about a specific user - Phase 6-2: Real data"""
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection not available")
-    
-    try:
-        with conn.cursor() as cur:
-            # Get user basic info
-            cur.execute("""
-                SELECT id, email, full_name, role, company_name, phone, 
-                       state, plan, is_active, stripe_customer_id,
-                       created_at, last_login, updated_at
-                FROM users 
-                WHERE id = %s
-            """, (user_id,))
-            
-            user_row = cur.fetchone()
-            if not user_row:
-                raise HTTPException(status_code=404, detail=f"User {user_id} not found")
-            
-            # Parse user data
-            user_id, email, full_name, role, company_name, phone, state, plan, is_active, stripe_customer_id, created_at, last_login, updated_at = user_row
-            
-            # Get deed statistics
-            cur.execute("""
-                SELECT 
-                    COUNT(*) as total_deeds,
-                    COUNT(*) FILTER (WHERE deed_type IS NOT NULL) as completed_deeds,
-                    0 as draft_deeds
-                FROM deeds 
-                WHERE user_id = %s AND COALESCE(status, '') <> 'deleted'
-            """, (user_id,))
-            
-            deed_stats = cur.fetchone()
-            total_deeds, completed_deeds, draft_deeds = deed_stats if deed_stats else (0, 0, 0)
-            
-            # Parse name (split full_name if available)
-            name_parts = full_name.split(' ', 1) if full_name else ['', '']
-            first_name = name_parts[0] if len(name_parts) > 0 else ''
-            last_name = name_parts[1] if len(name_parts) > 1 else ''
-            
-            # Build response with real data
-            user_details = {
-                "id": user_id,
-                "email": email or "",
-                "first_name": first_name,
-                "last_name": last_name,
-                "username": email.split('@')[0] if email else "",
-                "city": "",  # Not in current schema
-                "country": "USA",  # Default
-                "state": state or "",
-                "company_name": company_name or "",
-                "phone": phone or "",
-                "created_at": created_at.isoformat() if created_at else "",
-                "last_login": last_login.isoformat() if last_login else "",
-                "is_active": bool(is_active),
-                "role": role or "user",
-                "subscription_plan": plan or "free",
-                "subscription_status": "active" if is_active else "inactive",
-                "stripe_customer_id": stripe_customer_id or "",
-                "deed_statistics": {
-                    "total_deeds": total_deeds,
-                    "completed_deeds": completed_deeds,
-                    "draft_deeds": draft_deeds,
-                    "shared_deeds": 0,  # TODO: Implement when shared_deeds table exists
-                    "approved_deeds": completed_deeds
-                },
-                "activity_log": []  # TODO: Implement activity tracking table
-            }
-            
-            return user_details
-            
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching user details: {e}")
-        raise HTTPException(status_code=500, detail=f"Error fetching user details: {str(e)}")
-
-@app.put("/admin/users/{user_id}", dependencies=[Depends(get_current_admin)])
-def admin_update_user(user_id: int, user_update: AdminUserUpdate):
-    """Update user information (admin only)"""
-    # In production, update user in database
-    return {
-        "success": True,
-        "message": f"User {user_id} updated successfully",
-        "updated_fields": user_update.dict(exclude_unset=True)
-    }
-
-@app.delete("/admin/users/{user_id}", dependencies=[Depends(get_current_admin)])
-def admin_delete_user(user_id: int):
-    """Delete/deactivate a user (admin only)"""
-    # In production, soft delete or deactivate user
-    return {
-        "success": True,
-        "message": f"User {user_id} has been deactivated"
-    }
-
-@app.get("/admin/deeds", dependencies=[Depends(get_current_admin)])
-def admin_list_all_deeds(
-    page: int = Query(1, ge=1),
-    limit: int = Query(50, ge=1, le=100),
-    status: Optional[str] = None,
-    user_id: Optional[int] = None
-):
-    """List all deeds across all users"""
-    # Query real deed data from database
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            # Build query with filters
-            where_conditions = ["TRUE"]
-            params = []
-            
-            if user_id:
-                where_conditions.append("d.user_id = %s")
-                params.append(user_id)
-            
-            # Note: status filter can be added when status column is added to deeds table
-            
-            where_clause = " AND ".join(where_conditions)
-            
-            # Get total count for pagination
-            count_query = f"SELECT COUNT(*) FROM deeds d WHERE {where_clause}"
-            cur.execute(count_query, params)
-            total_count = cur.fetchone()[0]
-            
-            # Get deeds with user information
-            offset = (page - 1) * limit
-            query = f"""
-                SELECT d.id, d.user_id, d.deed_type, d.property_address, 
-                       d.grantor_name, d.grantee_name, d.created_at, d.updated_at,
-                       u.email, u.full_name
-                FROM deeds d
-                JOIN users u ON d.user_id = u.id
-                WHERE {where_clause}
-                ORDER BY d.created_at DESC
-                LIMIT %s OFFSET %s
-            """
-            params.extend([limit, offset])
-            cur.execute(query, params)
-            deeds = cur.fetchall()
-            
-            # Format deeds for response
-            formatted_deeds = []
-            for deed in deeds:
-                formatted_deeds.append({
-                    "id": deed[0],
-                    "user_id": deed[1],
-                    "deed_type": deed[2],
-                    "property_address": deed[3],
-                    "grantor_name": deed[4],
-                    "grantee_name": deed[5],
-                    "created_at": deed[6].isoformat() if deed[6] else None,
-                    "updated_at": deed[7].isoformat() if deed[7] else None,
-                    "user_email": deed[8],
-                    "user_name": deed[9],
-                    "status": "completed",  # Default - can add status column later
-                    "recorded": False,      # Default - can add recorded column later
-                    "shared_count": 0,      # Default - can add sharing tracking later
-                    "approval_count": 0     # Default - can add approval tracking later
-                })
-            
-            return {
-                "deeds": formatted_deeds,
-                "total": total_count,
-                "page": page,
-                "limit": limit,
-                "total_pages": (total_count + limit - 1) // limit
-            }
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching admin deeds: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch deeds: {str(e)}")
-
-@app.get("/admin/revenue", dependencies=[Depends(get_current_admin)])
-def admin_revenue_analytics():
-    """Get real revenue analytics from Stripe"""
-    try:
-        # Calculate time ranges
-        now = datetime.now()
-        start_of_month = datetime(now.year, now.month, 1)
-        start_of_month_ts = int(start_of_month.timestamp())
-        
-        # Get active subscriptions for MRR
-        subscriptions = stripe.Subscription.list(status="active", limit=100)
-        mrr_cents = sum(sub.plan.amount for sub in subscriptions.data if sub.plan)
-        
-        # Get charges this month
-        charges = stripe.Charge.list(created={"gte": start_of_month_ts}, limit=100)
-        monthly_revenue_cents = sum(c.amount for c in charges.data if c.paid and not c.refunded)
-        stripe_fees_cents = sum(c.balance_transaction.fee if hasattr(c, 'balance_transaction') and c.balance_transaction else 0 for c in charges.data if c.paid)
-        
-        # Get refunds this month
-        refunds = stripe.Refund.list(created={"gte": start_of_month_ts}, limit=100)
-        refunds_cents = sum(r.amount for r in refunds.data)
-        
-        # Get all-time charges for total revenue (last 12 months to keep it manageable)
-        twelve_months_ago = int((now - timedelta(days=365)).timestamp())
-        all_charges = stripe.Charge.list(created={"gte": twelve_months_ago}, limit=100)
-        total_revenue_cents = sum(c.amount for c in all_charges.data if c.paid and not c.refunded)
-        
-        # Build monthly breakdown (last 6 months)
-        monthly_breakdown = []
-        for i in range(6):
-            month_date = now - timedelta(days=30 * i)
-            month_str = month_date.strftime("%Y-%m")
-            month_start = datetime(month_date.year, month_date.month, 1)
-            month_end = (month_start + timedelta(days=32)).replace(day=1)
-            
-            month_charges = stripe.Charge.list(
-                created={"gte": int(month_start.timestamp()), "lt": int(month_end.timestamp())},
-                limit=100
-            )
-            revenue = sum(c.amount for c in month_charges.data if c.paid and not c.refunded)
-            
-            monthly_breakdown.append({
-                "month": month_str,
-                "revenue_cents": revenue,
-                "revenue_dollars": revenue / 100
-            })
-        
-        revenue_data = {
-            "overview": {
-                "total_revenue_cents": total_revenue_cents,
-                "monthly_revenue_cents": monthly_revenue_cents,
-                "stripe_fees_cents": stripe_fees_cents,
-                "refunds_cents": refunds_cents,
-                "net_monthly_revenue_cents": monthly_revenue_cents - stripe_fees_cents - refunds_cents
-            },
-            "monthly_breakdown": monthly_breakdown,
-            "mrr_arr": {
-                "mrr_cents": mrr_cents,
-                "mrr_dollars": mrr_cents / 100,
-                "arr_cents": mrr_cents * 12,
-                "arr_dollars": (mrr_cents * 12) / 100
-            }
-        }
-        
-        return revenue_data
-        
-    except stripe.error.StripeError as e:
-        print(f"Stripe error: {e}")
-        # Return zero values if Stripe fails
-        return {
-            "overview": {
-                "total_revenue_cents": 0,
-                "monthly_revenue_cents": 0,
-                "stripe_fees_cents": 0,
-                "refunds_cents": 0,
-                "net_monthly_revenue_cents": 0
-            },
-            "monthly_breakdown": [],
-            "mrr_arr": {
-                "mrr_cents": 0,
-                "mrr_dollars": 0,
-                "arr_cents": 0,
-                "arr_dollars": 0
-            }
-        }
-    except Exception as e:
-        print(f"Revenue analytics error: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch revenue data: {str(e)}")
-
-@app.get("/admin/analytics", dependencies=[Depends(get_current_admin)])
-def admin_platform_analytics():
-    """Get comprehensive platform analytics"""
-    analytics_data = {
-        "user_metrics": {
-            "total_users": 1247,
-            "active_users_30d": 892,
-            "new_users_7d": 23,
-            "user_retention_rate": 78.5,
-            "average_session_duration": "12:34"
-        },
-        "deed_metrics": {
-            "total_deeds": 3456,
-            "completed_deeds": 2891,
-            "draft_deeds": 565,
-            "completion_rate": 83.6,
-            "average_completion_time": "2.3 days"
-        },
-        "sharing_metrics": {
-            "total_shares": 1234,
-            "approval_rate": 76.8,
-            "average_response_time": "1.2 days",
-            "most_shared_deed_type": "Quitclaim Deed"
-        },
-        "geographic_distribution": [
-            {"state": "California", "user_count": 423, "deed_count": 1234},
-            {"state": "Texas", "user_count": 298, "deed_count": 876},
-            {"state": "Florida", "user_count": 187, "deed_count": 543}
-        ],
-        "performance_metrics": {
-            "api_response_time": "145ms",
-            "uptime": "99.8%",
-            "error_rate": "0.12%"
-        }
-    }
-    
-    return analytics_data
-
-@app.get("/admin/system-health", dependencies=[Depends(get_current_admin)])
-def admin_system_health():
-    """Get system health and status information"""
-    health_data = {
-        "status": "healthy",
-        "timestamp": datetime.now().isoformat(),
-        "services": {
-            "api": {"status": "up", "response_time": "145ms"},
-            "database": {"status": "up", "connections": 23, "max_connections": 100},
-            "stripe": {"status": "up", "last_webhook": "2024-01-15T09:45:00Z"},
-            "email": {"status": "up", "queue_size": 5}
-        },
-        "resources": {
-            "cpu_usage": "34%",
-            "memory_usage": "67%",
-            "disk_usage": "45%"
-        },
-        "recent_errors": [
-            {"timestamp": "2024-01-15T08:30:00Z", "level": "warning", "message": "High memory usage detected"},
-            {"timestamp": "2024-01-14T22:15:00Z", "level": "error", "message": "Payment webhook failed for user 456"}
-        ]
-    }
-    
-    return health_data
-
-
-@app.get("/admin/system/overview", dependencies=[Depends(get_current_admin)])
-def admin_system_overview():
-    """Get system overview with real health checks and PDF stats - Phase 5D"""
-    # Check database health
-    db_status = "down"
-    db_latency = 0
-    try:
-        import time
-        start = time.time()
-        with conn.cursor() as cur:
-            cur.execute("SELECT 1")
-            cur.fetchone()
-        db_latency = int((time.time() - start) * 1000)
-        db_status = "up"
-    except Exception as e:
-        print(f"DB health check failed: {e}")
-    
-    # Check Stripe health
-    stripe_status = "down"
-    try:
-        stripe.Account.retrieve()
-        stripe_status = "up"
-    except Exception as e:
-        print(f"Stripe health check failed: {e}")
-    
-    # PDF engine status
-    pdf_engine_primary = os.getenv("PDF_ENGINE", "auto")
-    use_pdfshift = os.getenv("USE_PDFSHIFT", "false").lower() == "true"
-    pdfshift_key = os.getenv("PDFSHIFT_API_KEY")
-    
-    if use_pdfshift and pdfshift_key:
-        pdf_primary = "PDFShift"
-    else:
-        pdf_primary = "WeasyPrint"
-    
-    # Get PDF generation stats from deeds table
-    pdf_stats = {
-        "total_generated": 0,
-        "pdfshift_count": 0,
-        "weasyprint_count": 0,
-        "avg_time_ms": 0,
-        "by_type": {}
-    }
-    
-    try:
-        with conn.cursor() as cur:
-            # Total deeds with PDFs (completed deeds)
-            cur.execute("SELECT COUNT(*) FROM deeds WHERE status = 'completed'")
-            pdf_stats["total_generated"] = cur.fetchone()[0] or 0
-            
-            # For now, estimate PDFShift vs WeasyPrint based on deployment date
-            # (In future, add pdf_engine column to deeds table)
-            if use_pdfshift:
-                pdf_stats["pdfshift_count"] = pdf_stats["total_generated"]
-                pdf_stats["weasyprint_count"] = 0
-            else:
-                pdf_stats["pdfshift_count"] = 0
-                pdf_stats["weasyprint_count"] = pdf_stats["total_generated"]
-            
-            # By deed type
-            cur.execute("""
-                SELECT deed_type, COUNT(*) as count 
-                FROM deeds 
-                WHERE status = 'completed'
-                GROUP BY deed_type
-                ORDER BY count DESC
-            """)
-            for row in cur.fetchall():
-                pdf_stats["by_type"][row[0]] = row[1]
-                
-    except Exception as e:
-        print(f"PDF stats error: {e}")
-    
-    return {
-        "health": {
-            "database": {"status": db_status, "latency_ms": db_latency},
-            "pdf_engine": {"status": "up", "primary": pdf_primary},
-            "sitex": {"status": "unknown", "last_call": None},
-            "stripe": {"status": stripe_status}
-        },
-        "pdf_stats": pdf_stats
-    }
-
-@app.get("/admin/system-metrics", dependencies=[Depends(get_current_admin)])
-def admin_system_metrics():
-    """Get real-time system metrics - Phase 6-2: Real monitoring data"""
-    reqs = METRICS.get('requests_total', 0)
-    lat_sum = METRICS.get('latency_ms_sum', 0)
-    avg_ms = int(lat_sum / reqs) if reqs else 0
-    
-    return {
-        "requests_total": reqs,
-        "avg_latency_ms": avg_ms,
-        "status": {k: v for k, v in METRICS.items() if str(k).startswith('status_')},
-        "last_request_unix": LAST_REQUEST_TS,
-        "last_request_iso": datetime.fromtimestamp(LAST_REQUEST_TS).isoformat() if LAST_REQUEST_TS > 0 else None
-    }
-
-# ============================================================================
-# USER ENDPOINTS - Regular User Operations
-# ============================================================================
-
-# T5: unauthenticated placeholder endpoints removed (payment-methods,
-# subscriptions, deed status/recipients, bare property search, inline
-# generate-deed and preview, POST /users, GET /users/{email}) — all
-# returned fake data and had zero frontend consumers.
-
-
-
-@app.get("/user/me")
-def get_current_user(user_id: int = Depends(get_current_user_id)):
-    """Get current user information from JWT token"""
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT id, email, full_name, role, company_name, phone, state, plan
-                FROM users 
-                WHERE id = %s AND is_active = TRUE
-            """, (user_id,))
-            
-            user = cur.fetchone()
-            
-            if not user:
-                raise HTTPException(status_code=404, detail="User not found")
-            
-            # Split full_name into first and last name
-            full_name = user[2] or ""
-            name_parts = full_name.split(" ", 1)
-            first_name = name_parts[0] if name_parts else ""
-            last_name = name_parts[1] if len(name_parts) > 1 else ""
-            
-            return {
-                "id": user[0],
-                "email": user[1],
-                "first_name": first_name,
-                "last_name": last_name,
-                "full_name": user[2],
-                "role": user[3],
-                "company": user[4],
-                "phone": user[5],
-                "state": user[6],
-                "plan": user[7]
-            }
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching current user: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch user: {str(e)}")
-
-# Deed endpoints
-@app.post("/deeds")
-def create_deed_endpoint(deed: DeedCreate, user_id: int = Depends(get_current_user_id)):
-    """Create a new deed with validation - Backend Hotfix V1"""
-    
-    # Convert Pydantic model to dict
-    deed_data = deed.dict()
-    
-    # DEFENSIVE: Strip whitespace and validate non-empty for critical fields
-    # This provides an additional layer of validation beyond Pydantic
-    critical_fields = {
-        'grantor_name': 'Grantor information',
-        'grantee_name': 'Grantee information',
-        'legal_description': 'Legal description'
-    }
-    
-    for field_name, field_label in critical_fields.items():
-        value = (deed_data.get(field_name) or "").strip()
-        deed_data[field_name] = value
-        if not value:
-            print(f"[Backend /deeds] ❌ VALIDATION ERROR: {field_label} is empty!")
-            print(f"[Backend /deeds] Received payload: {deed_data}")
-            raise HTTPException(
-                status_code=422, 
-                detail=f"Validation failed: {field_label} is required and cannot be empty"
-            )
-    
-    # Enhanced logging for diagnostics
-    print(f"[Backend /deeds] ✅ Creating deed for user_id={user_id}")
-    print(f"[Backend /deeds] deed_type: {deed_data.get('deed_type')}")
-    print(f"[Backend /deeds] county: {deed_data.get('county')}")  # ✅ PHASE 19: Add county logging
-    print(f"[Backend /deeds] grantor_name: {deed_data.get('grantor_name')}")
-    print(f"[Backend /deeds] grantee_name: {deed_data.get('grantee_name')}")
-    print(f"[Backend /deeds] legal_description: {deed_data.get('legal_description')[:100]}...")
-    print(f"[Backend /deeds] source: {deed_data.get('source', 'unknown')}")
-    
-    new_deed = create_deed(user_id, deed_data)
-
-    if not new_deed:
-        print(f"[Backend /deeds] ❌ create_deed returned None!")
-        raise HTTPException(status_code=500, detail="Failed to create deed - check backend logs")
-
-    # T2: render the deed's PDF once at generation time and store it.
-    # Non-blocking — if rendering fails the deed record still saves, and
-    # /download will retry the render on first request.
-    try:
-        from services.deed_pdf import generate_and_store
-        if conn:
-            digest = generate_and_store(conn, new_deed)
-            new_deed["pdf_url"] = f"/deeds/{new_deed['id']}/download"
-            print(f"[T2] Stored PDF for deed {new_deed['id']} (sha256 {digest[:12]}…)")
-    except Exception as pdf_error:
-        print(f"[T2] PDF generation failed for deed {new_deed.get('id')} (non-blocking): {pdf_error}")
-    
-    # Phase 7: Send deed completion notification
-    try:
-        from utils.notifications import send_deed_completion_notification
-        
-        # Get user details for notification
-        with conn.cursor() as cur:
-            cur.execute("SELECT email, full_name FROM users WHERE id = %s", (user_id,))
-            user_row = cur.fetchone()
-            
-            if user_row:
-                user_email = user_row[0]
-                user_name = user_row[1] or "Valued Customer"
-                deed_type = deed_data.get('deed_type', 'deed')
-                deed_id = new_deed.get('id', 0)
-                
-                # Send notification (non-blocking)
-                notification_sent = send_deed_completion_notification(
-                    user_email=user_email,
-                    user_name=user_name,
-                    deed_type=deed_type,
-                    deed_id=deed_id
-                )
-                
-                if notification_sent:
-                    print(f"[Phase 7] ✅ Deed completion email sent to {user_email}")
-                else:
-                    print(f"[Phase 7] ⚠️ Failed to send deed completion email")
-    except Exception as notif_error:
-        # Don't fail the request if notification fails
-        print(f"[Phase 7] ⚠️ Notification error (non-blocking): {notif_error}")
-    
-    return new_deed
-
-@app.get("/deeds")
-def list_deeds_endpoint(user_id: int = Depends(get_current_user_id)):
-    """List all deeds for current user"""
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT id, deed_type, property_address, grantor_name, grantee_name,
-                       created_at, updated_at
-                FROM deeds
-                WHERE user_id = %s AND COALESCE(status, '') <> 'deleted'
-                ORDER BY created_at DESC
-            """, (user_id,))
-            
-            deeds = cur.fetchall()
-            
-            # Format deeds for frontend (Phase 6-1: Fixed field names to match frontend)
-            formatted_deeds = []
-            for deed in deeds:
-                formatted_deeds.append({
-                    "id": deed[0],
-                    "deed_type": deed[1],
-                    "property": deed[2],  # Frontend expects 'property' not 'address'
-                    "grantor": deed[3],
-                    "grantee": deed[4],
-                    "created_at": deed[5].strftime("%Y-%m-%d") if deed[5] else None,  # Frontend expects 'created_at' not 'date'
-                    "updated_at": deed[6].strftime("%Y-%m-%d") if deed[6] else None,  # Add updated_at
-                    "status": "completed",  # Default status - you can add a status column later
-                    "recorded": False  # Default - you can add a recorded column later
-                })
-            
-            return {"deeds": formatted_deeds}
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching deeds: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch deeds: {str(e)}")
-
-# --- Phase 6-1: Deeds summary endpoint (for dashboard) ---
-@app.get("/deeds/summary")
-def deeds_summary(user_id: int = Depends(get_current_user_id)) -> Dict[str, int]:
-    """Return aggregated deed counts for the current user."""
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            # Query for status counts
-            cur.execute("""
-                SELECT 
-                    COUNT(*) as total,
-                    COUNT(*) FILTER (WHERE deed_type IS NOT NULL) as completed,
-                    0 as in_progress
-                FROM deeds 
-                WHERE user_id = %s AND COALESCE(status, '') <> 'deleted'
-            """, (user_id,))
-            
-            row = cur.fetchone()
-            total = row[0] if row else 0
-            completed = row[1] if row else 0
-            in_progress = row[2] if row else 0
-            
-            # Query for deeds created this month
-            cur.execute("""
-                SELECT COUNT(*) 
-                FROM deeds 
-                WHERE user_id = %s AND COALESCE(status, '') <> 'deleted' 
-                AND created_at >= date_trunc('month', CURRENT_DATE)
-            """, (user_id,))
-            
-            month_row = cur.fetchone()
-            month = month_row[0] if month_row else 0
-            
-            return {
-                "total": total,
-                "completed": completed,
-                "in_progress": in_progress,
-                "month": month
-            }
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching deed summary: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch summary: {str(e)}")
-
-# --- Phase 6-2: Wizard draft persistence (minimal in-memory) ---
-@app.post("/deeds/drafts")
-def save_draft(payload: DraftPayload, user_id: int = Depends(get_current_user_id)):
-    """Save a draft deed for the current user"""
-    draft_id = f"{user_id}:{int(time())}"
-    if user_id not in _DRAFTS:
-        _DRAFTS[user_id] = {}
-    _DRAFTS[user_id][draft_id] = {
-        "deed_type": payload.deed_type,
-        "data": payload.data,
-        "updated_at": int(time())
-    }
-    return {"draft_id": draft_id, "message": "Draft saved successfully"}
-
-@app.get("/deeds/drafts")
-def list_drafts(user_id: int = Depends(get_current_user_id)):
-    """List all draft deeds for the current user"""
-    user_drafts = _DRAFTS.get(user_id, {})
-    return [
-        {"id": draft_id, **draft_data}
-        for draft_id, draft_data in user_drafts.items()
-    ]
-
-@app.get("/deeds/available")
-def list_available_deeds_for_sharing(user_id: int = Depends(get_current_user_id)):
-    """List deeds available for sharing (completed deeds)"""
-    try:
-        if not conn:
-            raise HTTPException(status_code=500, detail="Database connection not available")
-        
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT id, property_address, deed_type
-                FROM deeds 
-                WHERE user_id = %s AND COALESCE(status, '') <> 'deleted' 
-                ORDER BY created_at DESC
-            """, (user_id,))
-            
-            deeds = cur.fetchall()
-            
-            # Format for sharing dropdown
-            available_deeds = []
-            for deed in deeds:
-                available_deeds.append({
-                    "id": deed[0],
-                    "address": deed[1],
-                    "deed_type": deed[2]
-                })
-            
-            return {"available_deeds": available_deeds}
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching available deeds: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch available deeds: {str(e)}")
-
-@app.get("/deeds/{deed_id}")
-def get_deed_endpoint(deed_id: int, user_id: int = Depends(get_current_user_id)):
-    """Get a specific deed - Phase 15 v5: Preview page integration"""
-    try:
-        cursor = conn.cursor(cursor_factory=RealDictCursor)
-        
-        # Fetch deed - ensure user owns it or is admin
-        cursor.execute("""
-            SELECT d.*, u.role 
-            FROM deeds d
-            LEFT JOIN users u ON u.id = %s
-            WHERE d.id = %s AND (d.user_id = %s OR u.role = 'admin')
-        """, (user_id, deed_id, user_id))
-        
-        deed = cursor.fetchone()
-        cursor.close()
-        
-        if not deed:
-            raise HTTPException(status_code=404, detail="Deed not found")
-        
-        return dict(deed)
-    except HTTPException:
-        raise
-    except Exception as e:
-        print(f"Error fetching deed {deed_id}: {e}")
-        conn.rollback()
-        raise HTTPException(status_code=500, detail=f"Failed to fetch deed: {str(e)}")
-
-
-# Shared Deeds endpoints
-@app.post("/shared-deeds")
-def share_deed_for_approval(share_data: ShareDeedCreate, user_id: int = Depends(get_current_user_id)):
-    """Share a deed with someone for approval - Enhanced with configurable expiration"""
-    import uuid
-    from datetime import timezone
-    # Use configurable expiration (default 7 days = 168 hours)
-    expires_in = share_data.expires_in_hours or 168
-    expires_at = datetime.now(timezone.utc) + timedelta(hours=expires_in)
-    approval_token = str(uuid.uuid4())
-    
-    # Get the owner's name and deed details from database
-    owner_name = "DeedPro User"  # Default
-    deed_type = "deed"  # Default
-    
-    try:
-        with conn.cursor() as cur:
-            # Get owner's name
-            cur.execute("SELECT full_name FROM users WHERE id = %s", (user_id,))
-            owner_row = cur.fetchone()
-            if owner_row:
-                owner_name = owner_row[0] or owner_name
-            
-            # Get deed type
-            cur.execute("SELECT deed_type FROM deeds WHERE id = %s", (share_data.deed_id,))
-            deed_row = cur.fetchone()
-            if deed_row:
-                deed_type = deed_row[0] or deed_type
-    except Exception as db_error:
-        print(f"[Phase 7] Warning: Could not fetch owner/deed details: {db_error}")
-    
-    # Generate approval URL
-    app_url = os.getenv('FRONTEND_URL', 'https://deedpro-frontend-new.vercel.app')
-    approval_url = f"{app_url}/approve/{approval_token}"
-    
-    # Phase 7.5: Save to deed_shares table (gap-plan fix)
-    shared_deed_id = None
-    property_address = "Unknown"
-    apn = "Unknown"
-    
-    try:
-        with conn.cursor() as cur:
-            # Get deed details
-            cur.execute("""
-                SELECT property_address, apn 
-                FROM deeds 
-                WHERE id = %s
-            """, (share_data.deed_id,))
-            deed_info = cur.fetchone()
-            if deed_info:
-                property_address = deed_info[0] or property_address
-                apn = deed_info[1] or apn
-            
-            # Insert into deed_shares table
-            cur.execute("""
-                INSERT INTO deed_shares (
-                    deed_id, owner_user_id, recipient_email, token, 
-                    status, expires_at, created_at, updated_at
-                )
-                VALUES (%s, %s, %s, %s, %s, %s, NOW(), NOW())
-                RETURNING id
-            """, (
-                share_data.deed_id,
-                user_id,
-                share_data.recipient_email,
-                approval_token,
-                'sent',
-                expires_at
-            ))
-            result = cur.fetchone()
-            if result:
-                shared_deed_id = result.get('id') if isinstance(result, dict) else result[0]
-            conn.commit()
-            print(f"[Phase 7.5] ✅ Share saved to database: ID {shared_deed_id}")
-    except Exception as db_error:
-        conn.rollback()  # Phase 7.5 FIX: Rollback transaction on error
-        print(f"[Phase 7.5] ⚠️ Failed to save share to database: {db_error}")
-        # Continue anyway - at least send the email
-    
-    shared_deed = {
-        "id": shared_deed_id or 101,  # Use real ID from database
-        "deed_id": share_data.deed_id,
-        "shared_by_user_id": user_id,
-        "recipient_name": share_data.recipient_name,
-        "recipient_email": share_data.recipient_email,
-        "recipient_role": share_data.recipient_role,
-        "message": share_data.message,
-        "status": "sent",
-        "approval_token": approval_token,
-        "expires_at": expires_at.isoformat(),
-        "created_at": datetime.now().isoformat(),
-        "approval_url": approval_url,
-        "property_address": property_address,
-        "apn": apn
-    }
-    
-    # Phase 7: Send sharing notification email
-    email_sent = False
-    try:
-        from utils.notifications import send_share_notification
-        
-        email_sent = send_share_notification(
-            recipient_email=share_data.recipient_email,
-            recipient_name=share_data.recipient_name,
-            owner_name=owner_name,
-            deed_type=deed_type,
-            share_link=approval_url
-        )
-        
-        if email_sent:
-            print(f"[Phase 7] ✅ Sharing notification sent to {share_data.recipient_email}")
-        else:
-            print(f"[Phase 7] ⚠️ Failed to send sharing notification")
-    except Exception as notif_error:
-        # Don't fail the request if notification fails
-        print(f"[Phase 7] ⚠️ Sharing notification error (non-blocking): {notif_error}")
-    
-    if not email_sent:
-        # Still return success but indicate email issue
-        print("[Phase 7] Warning: Deed shared but email notification failed")
-    
-    return {
-        "success": True,
-        "message": "Deed shared successfully! Approval email sent." if email_sent else "Deed shared, but email notification failed.",
-        "shared_deed": shared_deed,
-        "email_sent": email_sent
-    }
-
-@app.get("/shared-deeds")
-def list_shared_deeds(user_id: int = Depends(get_current_user_id)):
-    """List all shared deeds for current user - Phase 7.5: Read from deed_shares table"""
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection not available")
-    
-    try:
-        with conn.cursor() as cur:
-            # Phase 7.5: Query from deed_shares table (gap-plan schema)
-            # Show deeds shared BY this user (as owner)
-            cur.execute("""
-                SELECT 
-                    ds.id,
-                    ds.deed_id,
-                    ds.owner_user_id,
-                    ds.recipient_email,
-                    ds.status,
-                    ds.token,
-                    ds.expires_at,
-                    ds.created_at,
-                    ds.updated_at,
-                    d.property_address,
-                    d.deed_type,
-                    u.full_name as owner_name
-                FROM deed_shares ds
-                JOIN deeds d ON ds.deed_id = d.id
-                JOIN users u ON ds.owner_user_id = u.id
-                WHERE ds.owner_user_id = %s
-                ORDER BY ds.created_at DESC
-            """, (user_id,))
-            
-            rows = cur.fetchall()
-            
-            # Phase 7.5 FIX: Map deed_shares columns (RealDictCursor returns dicts)
-            shared_deeds = []
-            for row in rows:
-                # RealDictCursor returns dictionaries, not tuples
-                expires_at = row.get('expires_at') if isinstance(row, dict) else row[6]
-                created_at = row.get('created_at') if isinstance(row, dict) else row[7]
-                updated_at = row.get('updated_at') if isinstance(row, dict) else row[8]
-                
-                shared_deeds.append({
-                    "id": row.get('id') if isinstance(row, dict) else row[0],
-                    "deed_id": row.get('deed_id') if isinstance(row, dict) else row[1],
-                    "shared_by_id": row.get('owner_user_id') if isinstance(row, dict) else row[2],
-                    "shared_with_email": row.get('recipient_email') if isinstance(row, dict) else row[3],
-                    "status": row.get('status') if isinstance(row, dict) else row[4],
-                    "message": f"Shared via link - expires {expires_at.strftime('%Y-%m-%d') if expires_at else 'never'}",
-                    "share_type": "review",
-                    "date": created_at.isoformat() if created_at else "",
-                    "updated_at": updated_at.isoformat() if updated_at else "",
-                    "property": (row.get('property_address') if isinstance(row, dict) else row[9]) or "",
-                    "type": (row.get('deed_type') if isinstance(row, dict) else row[10]) or "",
-                    "shared_by": (row.get('owner_name') if isinstance(row, dict) else row[11]) or "Unknown User",
-                    "approval_token": row.get('token') if isinstance(row, dict) else row[5]
-                })
-            
-            print(f"[Phase 7.5] ✅ Fetched {len(shared_deeds)} shared deeds for user {user_id}")
-            return shared_deeds
-            
-    except Exception as e:
-        conn.rollback()  # Phase 14-B: Prevent transaction cascade failures
-        print(f"Error fetching shared deeds: {e}")
-        # Graceful degradation: return empty array if table doesn't exist yet
-        if "does not exist" in str(e):
-            return []
-        raise HTTPException(status_code=500, detail=f"Failed to fetch shared deeds: {str(e)}")
-
-@app.post("/shared-deeds/{shared_deed_id}/resend")
-def resend_approval_email(shared_deed_id: int, user_id: int = Depends(get_current_user_id)):
-    """Resend the approval reminder email with expiration extension"""
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection not available")
-    
-    try:
-        from datetime import timezone
-        from utils.notifications import send_email
-        
-        with conn.cursor() as cur:
-            # Get the share details
-            cur.execute("""
-                SELECT ds.id, ds.token, ds.recipient_email, ds.status, ds.expires_at,
-                       d.deed_type, d.property_address, d.apn,
-                       u.email as owner_email, u.full_name as owner_name
-                FROM deed_shares ds
-                JOIN deeds d ON ds.deed_id = d.id
-                JOIN users u ON ds.owner_user_id = u.id
-                WHERE ds.id = %s AND ds.owner_user_id = %s
-            """, (shared_deed_id, user_id))
-            
-            row = cur.fetchone()
-            if not row:
-                raise HTTPException(status_code=404, detail="Share not found")
-            
-            # Extract data
-            if isinstance(row, dict):
-                share_id = row.get('id')
-                token = row.get('token')
-                recipient_email = row.get('recipient_email')
-                status = row.get('status')
-                expires_at = row.get('expires_at')
-                deed_type = row.get('deed_type')
-                property_address = row.get('property_address')
-                owner_name = row.get('owner_name') or 'DeedPro User'
-            else:
-                share_id, token, recipient_email, status, expires_at = row[0], row[1], row[2], row[3], row[4]
-                deed_type, property_address = row[5], row[6]
-                owner_name = row[9] or 'DeedPro User'
-            
-            # Check if can be resent
-            if status in ('approved', 'rejected', 'revoked'):
-                raise HTTPException(status_code=400, detail=f"Cannot resend - share is {status}")
-            
-            # Check if expired and extend expiration by 24 hours
-            now = datetime.now(timezone.utc)
-            if expires_at < now:
-                new_expiry = now + timedelta(hours=24)
-                cur.execute("""
-                    UPDATE deed_shares 
-                    SET expires_at = %s, updated_at = NOW()
-                    WHERE id = %s
-                """, (new_expiry, share_id))
-                expires_at = new_expiry
-            
-            # Build approval URL
-            app_url = os.getenv('FRONTEND_URL', 'https://deedpro-frontend-new.vercel.app')
-            approval_url = f"{app_url}/approve/{token}"
-            
-            # Calculate hours remaining
-            hours_remaining = max(0, int((expires_at - now).total_seconds() / 3600))
-            
-            # Send reminder email
-            subject = f"Reminder: Deed Review Pending - {property_address}"
-            html_content = f"""
-            <div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-                <h2 style="color: #1a1a1a;">Review Reminder</h2>
-                <p>This is a reminder that <strong>{owner_name}</strong> is waiting for your review of a deed:</p>
-                
-                <div style="background: #f5f5f5; padding: 20px; border-radius: 8px; margin: 20px 0;">
-                    <p style="margin: 5px 0;"><strong>Deed Type:</strong> {deed_type}</p>
-                    <p style="margin: 5px 0;"><strong>Property:</strong> {property_address}</p>
-                    <p style="margin: 5px 0; color: #e53e3e;"><strong>Expires in:</strong> {hours_remaining} hours</p>
-                </div>
-                
-                <a href="{approval_url}" 
-                   style="display: inline-block; background: #7C4DFF; color: white; padding: 12px 24px; 
-                          text-decoration: none; border-radius: 6px; font-weight: bold;">
-                    Review Deed Now
-                </a>
-                
-                <p style="color: #666; font-size: 14px; margin-top: 30px;">
-                    If you did not expect this email, you can safely ignore it.
-                </p>
-            </div>
-            """
-            
-            email_sent = send_email(recipient_email, subject, html_content)
-            
-            conn.commit()
-            
-            print(f"[Sharing] ✅ Reminder sent for share {share_id}, email_sent={email_sent}")
-            
-            return {
-                "success": True, 
-                "message": "Reminder sent",
-                "email_sent": email_sent,
-                "expires_at": expires_at.isoformat()
-            }
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()
-        print(f"[Sharing] ❌ Resend error: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-def _revoke_deed_share(shared_deed_id: int, user_id: int):
-    """Revoke a share in deed_shares (the live sharing table) with owner check.
-
-    T3 fix: the previous implementation targeted the legacy shared_deeds
-    table, which the live sharing stack never writes — revoke silently
-    no-op'd. Every other sharing endpoint uses deed_shares.
-    """
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection not available")
-    try:
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT owner_user_id FROM deed_shares
-                WHERE id = %s
-            """, (shared_deed_id,))
-            row = cur.fetchone()
-            if not row:
-                raise HTTPException(status_code=404, detail="Shared deed not found")
-            if row[0] != user_id:
-                raise HTTPException(status_code=403, detail="You don't have permission to revoke this share")
-
-            cur.execute("""
-                UPDATE deed_shares
-                SET status = 'revoked', updated_at = CURRENT_TIMESTAMP
-                WHERE id = %s
-            """, (shared_deed_id,))
-            conn.commit()
-
-        return {
-            "success": True,
-            "message": f"Access to shared deed {shared_deed_id} has been revoked"
-        }
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()
-        print(f"Error revoking shared deed: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to revoke shared deed: {str(e)}")
-
-@app.delete("/shared-deeds/{shared_deed_id}")
-def revoke_shared_deed(shared_deed_id: int, user_id: int = Depends(get_current_user_id)):
-    """Revoke access to a shared deed."""
-    return _revoke_deed_share(shared_deed_id, user_id)
-
-@app.post("/shared-deeds/{shared_deed_id}/revoke")
-def revoke_shared_deed_post(shared_deed_id: int, user_id: int = Depends(get_current_user_id)):
-    """Revoke access to a shared deed (the route the Shared Deeds UI calls)."""
-    return _revoke_deed_share(shared_deed_id, user_id)
-
-@app.get("/shared-deeds/{shared_deed_id}/feedback")
-def get_shared_deed_feedback(shared_deed_id: int, user_id: int = Depends(get_current_user_id)):
-    """Return recipient feedback for a share (owner-only). Called by the
-    Shared Deeds UI; feedback is written by the public rejection flow."""
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection not available")
-    try:
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT feedback, feedback_at, feedback_by, owner_user_id
-                FROM deed_shares
-                WHERE id = %s
-            """, (shared_deed_id,))
-            row = cur.fetchone()
-        if not row:
-            raise HTTPException(status_code=404, detail="Shared deed not found")
-        feedback, feedback_at, feedback_by, owner_user_id = row
-        if owner_user_id != user_id:
-            raise HTTPException(status_code=403, detail="You don't have permission to view this feedback")
-        return {
-            "feedback": feedback or "",
-            "feedback_at": feedback_at.isoformat() if feedback_at else None,
-            "feedback_by": feedback_by,
-        }
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()
-        print(f"Error fetching share feedback: {e}")
-        raise HTTPException(status_code=500, detail="Failed to fetch feedback")
-
-# Public approval endpoint (for recipients)
-@app.get("/approve/{approval_token}")
-def view_shared_deed(approval_token: str):
-    """Public endpoint for recipients to view shared deed - Enhanced with view tracking"""
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection unavailable")
-    
-    try:
-        from datetime import datetime, timezone
-        
-        with conn.cursor() as cur:
-            # Get share details with deed info (including county)
-            cur.execute("""
-                SELECT 
-                    ds.id, ds.deed_id, ds.owner_user_id, ds.recipient_email,
-                    ds.status, ds.expires_at, ds.created_at, ds.viewed_at,
-                    d.deed_type, d.property_address, d.apn, d.grantor_name, d.grantee_name,
-                    d.county, d.pdf_url,
-                    u.full_name as owner_name, u.email as owner_email
-                FROM deed_shares ds
-                JOIN deeds d ON d.id = ds.deed_id
-                LEFT JOIN users u ON u.id = ds.owner_user_id
-                WHERE ds.token = %s
-            """, (approval_token,))
-            
-            share = cur.fetchone()
-            
-            if not share:
-                raise HTTPException(status_code=404, detail="This approval link is invalid or does not exist")
-            
-            # Extract data (handle both dict and tuple)
-            if isinstance(share, dict):
-                share_id = share.get('id')
-                deed_id = share.get('deed_id')
-                owner_id = share.get('owner_user_id')
-                recipient_email = share.get('recipient_email')
-                status = share.get('status')
-                expires_at = share.get('expires_at')
-                viewed_at = share.get('viewed_at')
-                deed_type = share.get('deed_type')
-                property_address = share.get('property_address')
-                apn = share.get('apn')
-                grantor_name = share.get('grantor_name')
-                grantee_name = share.get('grantee_name')
-                county = share.get('county')
-                pdf_url = share.get('pdf_url')
-                owner_name = share.get('owner_name') or "DeedPro User"
-            else:
-                share_id = share[0]
-                deed_id = share[1]
-                owner_id = share[2]
-                recipient_email = share[3]
-                status = share[4]
-                expires_at = share[5]
-                viewed_at = share[7] if len(share) > 7 else None
-                deed_type = share[8] if len(share) > 8 else None
-                property_address = share[9] if len(share) > 9 else None
-                apn = share[10] if len(share) > 10 else None
-                grantor_name = share[11] if len(share) > 11 else None
-                grantee_name = share[12] if len(share) > 12 else None
-                county = share[13] if len(share) > 13 else None
-                pdf_url = share[14] if len(share) > 14 else None
-                owner_name = share[15] if len(share) > 15 else "DeedPro User"
-            
-            # Check if expired
-            now = datetime.now(timezone.utc)
-            is_expired = expires_at < now if expires_at else False
-            
-            # Update status to expired if needed
-            if is_expired and status == 'sent':
-                cur.execute("UPDATE deed_shares SET status = 'expired', updated_at = NOW() WHERE id = %s", (share_id,))
-                conn.commit()
-                raise HTTPException(status_code=410, detail="This approval link has expired")
-            
-            # Track first view (only if status is still 'sent')
-            if status == 'sent' and not viewed_at:
-                try:
-                    cur.execute("""
-                        UPDATE deed_shares 
-                        SET status = 'viewed', viewed_at = NOW(), updated_at = NOW()
-                        WHERE id = %s AND status = 'sent'
-                    """, (share_id,))
-                    conn.commit()
-                    status = 'viewed'
-                    print(f"[Sharing] ✅ First view tracked for share {share_id}")
-                except Exception as view_err:
-                    # Non-blocking - viewed_at column may not exist
-                    print(f"[Sharing] ⚠️ Could not track view: {view_err}")
-            
-            can_approve = not is_expired and status in ('sent', 'viewed')
-            
-            deed_data = {
-                "deed_id": deed_id,
-                "deed_type": deed_type,
-                "property_address": property_address,
-                "apn": apn,
-                "county": county,
-                "grantors": grantor_name,
-                "grantees": grantee_name,
-                "owner_name": owner_name,
-                "message": f"Please review this {deed_type}",
-                "expires_at": expires_at.isoformat() if expires_at else None,
-                "can_approve": can_approve,
-                "status": status,
-                "pdf_url": pdf_url,
-            }
-            
-            print(f"[Sharing] ✅ Approval link accessed: share_id={share_id}, status={status}, can_approve={can_approve}")
-            
-            return deed_data
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()
-        print(f"[Sharing] ❌ Error loading shared deed: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to load shared deed: {str(e)}")
-
-@app.post("/approve/{approval_token}")
-def submit_approval_response(approval_token: str, response: ApprovalResponse):
-    """Submit approval or rejection response - REJECTION BUNDLE: Enhanced with feedback, email, and notifications"""
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection unavailable")
-    
-    try:
-        from datetime import datetime, timezone
-        
-        with conn.cursor() as cur:
-            # Get share details WITH property address and owner info
-            cur.execute("""
-                SELECT ds.id, ds.status, ds.expires_at, ds.owner_user_id, ds.deed_id, 
-                       ds.recipient_email, d.property_address, u.email as owner_email
-                FROM deed_shares ds
-                JOIN deeds d ON d.id = ds.deed_id
-                LEFT JOIN users u ON u.id = ds.owner_user_id
-                WHERE ds.token = %s
-            """, (approval_token,))
-            
-            share = cur.fetchone()
-            
-            if not share:
-                raise HTTPException(status_code=404, detail="Invalid approval link")
-            
-            # Extract data (handle both dict and tuple)
-            if isinstance(share, dict):
-                share_id = share.get('id')
-                current_status = share.get('status')
-                expires_at = share.get('expires_at')
-                owner_id = share.get('owner_user_id')
-                deed_id = share.get('deed_id')
-                recipient_email = share.get('recipient_email')
-                property_address = share.get('property_address')
-                owner_email = share.get('owner_email')
-            else:
-                share_id = share[0]
-                current_status = share[1]
-                expires_at = share[2]
-                owner_id = share[3]
-                deed_id = share[4]
-                recipient_email = share[5]
-                property_address = share[6]
-                owner_email = share[7]
-            
-            # Check if expired
-            now = datetime.now(timezone.utc)
-            is_expired = expires_at < now if expires_at else False
-            
-            if is_expired:
-                raise HTTPException(status_code=410, detail="This approval link has expired")
-            
-            # Allow approval from 'sent' or 'viewed' status
-            if current_status not in ('sent', 'viewed'):
-                raise HTTPException(status_code=409, detail=f"This deed has already been {current_status}")
-            
-            # Get additional details for email (deed_type)
-            cur.execute("SELECT deed_type FROM deeds WHERE id = %s", (deed_id,))
-            deed_row = cur.fetchone()
-            deed_type = deed_row[0] if deed_row else "deed"
-            
-            # APPROVAL PATH - Enhanced with email notification
-            if response.approved:
-                cur.execute("""
-                    UPDATE deed_shares
-                    SET status = 'approved', updated_at = NOW()
-                    WHERE id = %s
-                """, (share_id,))
-                conn.commit()
-                
-                print(f"[Sharing] ✅ Deed approved: share_id={share_id}")
-                
-                # Send approval notification to owner
-                try:
-                    from utils.notifications import send_email
-                    
-                    if owner_email:
-                        app_url = os.getenv('FRONTEND_URL', 'https://deedpro-frontend-new.vercel.app')
-                        view_link = f"{app_url}/past-deeds"
-                        
-                        subject = f"✓ Deed Approved - {property_address}"
-                        html_content = f"""
-                        <div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-                            <div style="background: #10B981; color: white; padding: 20px; border-radius: 8px 8px 0 0; text-align: center;">
-                                <h2 style="margin: 0;">✓ Deed Approved</h2>
-                            </div>
-                            
-                            <div style="padding: 20px; background: #f9fafb; border-radius: 0 0 8px 8px;">
-                                <p>Great news! Your deed has been approved:</p>
-                                
-                                <div style="background: white; padding: 15px; border-radius: 8px; margin: 20px 0;">
-                                    <p style="margin: 5px 0;"><strong>Deed Type:</strong> {deed_type}</p>
-                                    <p style="margin: 5px 0;"><strong>Property:</strong> {property_address}</p>
-                                    <p style="margin: 5px 0;"><strong>Approved by:</strong> {recipient_email}</p>
-                                </div>
-                                
-                                <p>The deed is ready for recording.</p>
-                                
-                                <a href="{view_link}" 
-                                   style="display: inline-block; background: #7C4DFF; color: white; padding: 12px 24px; 
-                                          text-decoration: none; border-radius: 6px; font-weight: bold;">
-                                    View Deed
-                                </a>
-                            </div>
-                        </div>
-                        """
-                        
-                        email_sent = send_email(owner_email, subject, html_content)
-                        if email_sent:
-                            print(f"[Sharing] ✅ Approval email sent to {owner_email}")
-                except Exception as email_err:
-                    print(f"[Sharing] ⚠️ Approval email error (non-blocking): {email_err}")
-                
-                return {
-                    "success": True,
-                    "message": "Thank you! Your approval has been recorded. The owner has been notified.",
-                    "status": "approved"
-                }
-            
-            # REJECTION PATH (NEW: save feedback, email owner, create notification)
-            comments = (response.comments or "").strip()
-            
-            # 1. Save feedback to database
-            cur.execute("""
-                UPDATE deed_shares
-                SET status = 'rejected',
-                    feedback = %s,
-                    feedback_at = NOW(),
-                    feedback_by = %s,
-                    updated_at = NOW()
-                WHERE id = %s
-            """, (comments, recipient_email, share_id))
-            conn.commit()
-            
-            print(f"[REJECTION BUNDLE] ✅ Feedback saved: share_id={share_id}, length={len(comments)}")
-            
-            # 2. Send email notification to owner
-            try:
-                from utils.notifications import send_email, render_rejection_email
-                
-                if owner_email:
-                    link = f"{os.getenv('FRONTEND_URL', 'https://deedpro-frontend-new.vercel.app')}/shared-deeds?focus={share_id}"
-                    subject = "🔄 Deed Changes Requested - DeedPro"
-                    html = render_rejection_email(
-                        property_address or "your property",
-                        recipient_email or "Reviewer",
-                        comments,
-                        link
-                    )
-                    email_sent = send_email(owner_email, subject, html)
-                    
-                    if email_sent:
-                        print(f"[REJECTION BUNDLE] ✅ Email sent to owner: {owner_email}")
-                    else:
-                        print(f"[REJECTION BUNDLE] ⚠️ Email send failed (non-blocking)")
-            except Exception as email_error:
-                # Don't fail the request if email fails
-                print(f"[REJECTION BUNDLE] ⚠️ Email error (non-blocking): {email_error}")
-            
-            # 3. Create in-app notification (best-effort)
-            try:
-                from utils.notifications import create_notification
-                
-                notification_id = create_notification(
-                    conn,
-                    user_id=owner_id,
-                    ntype="share_rejected",
-                    title="📝 Changes Requested",
-                    message=f"{recipient_email or 'A reviewer'} requested changes for {property_address or 'your deed'}",
-                    link=f"/shared-deeds?focus={share_id}"
-                )
-                print(f"[REJECTION BUNDLE] ✅ Notification created: ID {notification_id}")
-            except Exception as notif_error:
-                # Rollback to prevent transaction cascade failure
-                conn.rollback()
-                # Notifications table may not exist or be incompatible
-                print(f"[REJECTION BUNDLE] ⚠️ Notification error (non-blocking): {notif_error}")
-            
-            return {
-                "success": True,
-                "message": "Thank you! Your feedback has been submitted. The owner has been notified.",
-                "status": "rejected",
-                "comments": comments
-            }
-            
-    except HTTPException:
-        raise
-    except Exception as e:
-        print(f"[REJECTION BUNDLE] ❌ Error: {e}")
-        import traceback
-        traceback.print_exc()
-        raise HTTPException(status_code=500, detail=f"Failed to submit approval: {str(e)}")
-
-# PDF access for shared deed recipients
-@app.get("/approve/{approval_token}/pdf")
-def get_shared_deed_pdf(approval_token: str):
-    """
-    Get the PDF for a shared deed.
-    Requires valid, non-expired token.
-    """
-    from fastapi.responses import Response
-    from datetime import timezone
-    
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection unavailable")
-    
-    try:
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT ds.status, ds.expires_at,
-                       d.pdf_data, d.pdf_url, d.deed_type, d.property_address
-                FROM deed_shares ds
-                JOIN deeds d ON ds.deed_id = d.id
-                WHERE ds.token = %s
-            """, (approval_token,))
-            
-            row = cur.fetchone()
-            if not row:
-                raise HTTPException(status_code=404, detail="Share not found")
-            
-            # Extract data
-            if isinstance(row, dict):
-                status = row.get('status')
-                expires_at = row.get('expires_at')
-                pdf_data = row.get('pdf_data')
-                pdf_url = row.get('pdf_url')
-                deed_type = row.get('deed_type', 'deed')
-                property_address = row.get('property_address', 'property')
-            else:
-                status, expires_at, pdf_data, pdf_url, deed_type, property_address = row[0], row[1], row[2], row[3], row[4] or 'deed', row[5] or 'property'
-            
-            # Check expiration
-            now = datetime.now(timezone.utc)
-            if expires_at and expires_at < now:
-                raise HTTPException(status_code=410, detail="This link has expired")
-            
-            # Check status (allow viewing even if rejected, but not if revoked)
-            if status == 'revoked':
-                raise HTTPException(status_code=403, detail="Access has been revoked")
-            
-            # Return PDF
-            if pdf_data:
-                # Clean filename
-                safe_filename = f"{deed_type}_{property_address}".replace(' ', '_').replace('/', '-')[:50]
-                return Response(
-                    content=pdf_data,
-                    media_type="application/pdf",
-                    headers={
-                        "Content-Disposition": f'inline; filename="{safe_filename}.pdf"'
-                    }
-                )
-            elif pdf_url:
-                # Redirect to PDF URL if no embedded data
-                from fastapi.responses import RedirectResponse
-                return RedirectResponse(url=pdf_url)
-            else:
-                raise HTTPException(status_code=404, detail="PDF not available for this deed")
-                
-    except HTTPException:
-        raise
-    except Exception as e:
-        print(f"[Sharing] ❌ PDF access error: {e}")
-        raise HTTPException(status_code=500, detail="Failed to retrieve PDF")
-
-
-# Recipients endpoints (from previous implementation)
-
-
-@app.delete("/deeds/{deed_id}")
-def delete_deed_endpoint(deed_id: int, user_id: int = Depends(get_current_user_id)):
-    """Soft-delete a deed (owner only): status='deleted', hidden from lists,
-    row and stored PDF retained in the database (T5 owner decision)."""
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection not available")
-    try:
-        with conn.cursor() as cur:
-            cur.execute("SELECT user_id FROM deeds WHERE id = %s", (deed_id,))
-            row = cur.fetchone()
-            if not row:
-                raise HTTPException(status_code=404, detail="Deed not found")
-            if row[0] != user_id:
-                raise HTTPException(status_code=403, detail="You don't have permission to delete this deed")
-            cur.execute("""
-                UPDATE deeds SET status = 'deleted', updated_at = CURRENT_TIMESTAMP
-                WHERE id = %s
-            """, (deed_id,))
-            conn.commit()
-        return {"success": True, "message": f"Deed {deed_id} deleted"}
-    except HTTPException:
-        raise
-    except Exception as e:
-        conn.rollback()
-        print(f"Error deleting deed {deed_id}: {e}")
-        raise HTTPException(status_code=500, detail="Failed to delete deed")
-
-@app.get("/deeds/{deed_id}/download")
-def download_deed_endpoint(deed_id: int, user_id: int = Depends(get_current_user_id)):
-    """Stream the stored deed PDF; render+store on first request for legacy rows."""
-    if not conn:
-        raise HTTPException(status_code=500, detail="Database connection not available")
-    try:
-        cursor = conn.cursor(cursor_factory=RealDictCursor)
-        cursor.execute("""
-            SELECT d.*, u.role
-            FROM deeds d
-            LEFT JOIN users u ON u.id = %s
-            WHERE d.id = %s AND (d.user_id = %s OR u.role = 'admin')
-        """, (user_id, deed_id, user_id))
-        deed = cursor.fetchone()
-        cursor.close()
-
-        if not deed:
-            raise HTTPException(status_code=404, detail="Deed not found")
-
-        with conn.cursor() as cur:
-            cur.execute("SELECT pdf_data FROM deed_pdfs WHERE deed_id = %s", (deed_id,))
-            stored = cur.fetchone()
-
-        if stored and stored[0]:
-            pdf_bytes = bytes(stored[0])
-        else:
-            # Legacy row saved before the stored-PDF pipeline: render now,
-            # store, and serve — subsequent downloads hit the stored copy.
-            from services.deed_pdf import render_deed_pdf, store_deed_pdf
-            pdf_bytes = render_deed_pdf(dict(deed))
-            store_deed_pdf(conn, deed_id, pdf_bytes)
-
-        return Response(
-            content=pdf_bytes,
-            media_type="application/pdf",
-            headers={"Content-Disposition": f'attachment; filename="deed_{deed_id}.pdf"'},
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        print(f"Error downloading deed {deed_id}: {e}")
-        conn.rollback()
-        raise HTTPException(status_code=500, detail="Failed to generate deed PDF")
-
-# Payment endpoints
-
-
-
-# Subscription endpoints
-
-
-# Property search endpoint (placeholder)
-
-# Enhanced deed preview endpoint with AI suggestions
-
-# Optional: Full deed generation endpoint with PDF
-
-class PriceUpdate(BaseModel):
-    plan_name: str
-    price: float
-    features: List[str]
-
-class NewPlan(BaseModel):
-    plan_name: str
-    price: float
-    features: List[str]
-
-# Pricing Endpoints
-@app.get("/pricing")
-async def get_pricing():
-    """Get all pricing plans for the landing page"""
-    try:
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT plan_name, price, features, is_active 
-                FROM pricing 
-                WHERE is_active = TRUE 
-                ORDER BY price ASC
-            """)
-            rows = cur.fetchall()
-            
-            # Phase 7.5 FIX: Handle both RealDictCursor (dict) and regular cursor (tuple)
-            result = []
-            for row in rows:
-                if isinstance(row, dict):
-                    # RealDictCursor
-                    result.append({
-                        "name": row.get('plan_name'),
-                        "price": float(row.get('price', 0)),
-                        "features": row.get('features') if row.get('features') else [],
-                        "popular": row.get('plan_name') == "professional"
-                    })
-                else:
-                    # Regular cursor (tuple)
-                    result.append({
-                        "name": row[0],
-                        "price": float(row[1]),
-                        "features": row[2] if row[2] else [],
-                        "popular": row[0] == "professional"
-                    })
-            
-            return result
-    except Exception as e:
-        conn.rollback()  # CRITICAL: Rollback to prevent transaction cascade failures
-        print(f"[PRICING ERROR] {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to fetch pricing: {str(e)}")
-
-@app.get("/pricing/plans")
-async def get_pricing_plans():
-    """Get all pricing plans - alternative endpoint for consistency"""
-    return await get_pricing()
-
-@app.post("/admin/create-plan")
-async def create_plan(plan: NewPlan, admin: str = Depends(get_current_admin)):
-    """Create new Stripe product/price and save to database"""
-    try:
-        # Create Stripe product
-        product = stripe.Product.create(
-            name=plan.plan_name.capitalize(),
-            type="service",
-            description=f"{plan.plan_name.capitalize()} plan with {len(plan.features)} features"
-        )
-        
-        # Create Stripe price
-        price = stripe.Price.create(
-            product=product.id,
-            unit_amount=int(plan.price * 100),  # Convert to cents
-            currency="usd",
-            recurring={"interval": "month"},
-            nickname=plan.plan_name
-        )
-        
-        # Save to database
-        with conn.cursor() as cur:
-            cur.execute("""
-                INSERT INTO pricing (plan_name, price, stripe_product_id, stripe_price_id, features, last_synced) 
-                VALUES (%s, %s, %s, %s, %s::jsonb, CURRENT_TIMESTAMP)
-                ON CONFLICT (plan_name) DO UPDATE SET 
-                    price = EXCLUDED.price,
-                    stripe_price_id = EXCLUDED.stripe_price_id,
-                    stripe_product_id = EXCLUDED.stripe_product_id,
-                    features = EXCLUDED.features,
-                    last_synced = CURRENT_TIMESTAMP
-            """, (plan.plan_name, plan.price, product.id, price.id, plan.features))
-            conn.commit()
-            
-        return {
-            "status": "created",
-            "plan_name": plan.plan_name,
-            "stripe_price_id": price.id,
-            "stripe_product_id": product.id
-        }
-        
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Failed to create plan: {str(e)}")
-
-@app.post("/admin/sync-pricing")
-async def sync_pricing(admin: str = Depends(get_current_admin)):
-    """Sync pricing from Stripe to database"""
-    try:
-        # Get all active prices from Stripe
-        prices = stripe.Price.list(active=True, limit=100).data
-        
-        synced_count = 0
-        with conn.cursor() as cur:
-            for price in prices:
-                if price.nickname:  # Only sync prices with nicknames (our plans)
-                    # Get product details
-                    product = stripe.Product.retrieve(price.product)
-                    
-                    # Update database
-                    cur.execute("""
-                        UPDATE pricing 
-                        SET price = %s, 
-                            stripe_price_id = %s,
-                            stripe_product_id = %s,
-                            last_synced = CURRENT_TIMESTAMP
-                        WHERE plan_name = %s
-                    """, (
-                        price.unit_amount / 100,  # Convert from cents
-                        price.id,
-                        product.id,
-                        price.nickname.lower()
-                    ))
-                    
-                    if cur.rowcount > 0:
-                        synced_count += 1
-                        
-            conn.commit()
-            
-        return {
-            "status": "synced", 
-            "synced_count": synced_count,
-            "message": f"Successfully synced {synced_count} plans from Stripe"
-        }
-        
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Failed to sync pricing: {str(e)}")
-
-@app.post("/admin/update-price")
-async def update_price(update: PriceUpdate, admin: str = Depends(get_current_admin)):
-    """Update price in both Stripe and database"""
-    try:
-        # Get current Stripe price ID from database
-        with conn.cursor() as cur:
-            cur.execute("""
-                SELECT stripe_price_id, stripe_product_id 
-                FROM pricing 
-                WHERE plan_name = %s
-            """, (update.plan_name,))
-            
-            result = cur.fetchone()
-            if not result:
-                raise HTTPException(status_code=404, detail="Plan not found")
-                
-            old_price_id, product_id = result
-        
-        # Create new price in Stripe (prices are immutable)
-        new_price = stripe.Price.create(
-            product=product_id,
-            unit_amount=int(update.price * 100),  # Convert to cents
-            currency="usd",
-            recurring={"interval": "month"},
-            nickname=update.plan_name
-        )
-        
-        # Deactivate old price
-        if old_price_id:
-            stripe.Price.modify(old_price_id, active=False)
-        
-        # Update database
-        with conn.cursor() as cur:
-            cur.execute("""
-                UPDATE pricing 
-                SET price = %s, 
-                    features = %s::jsonb,
-                    stripe_price_id = %s,
-                    last_synced = CURRENT_TIMESTAMP
-                WHERE plan_name = %s
-            """, (update.price, update.features, new_price.id, update.plan_name))
-            conn.commit()
-            
-        return {
-            "status": "updated",
-            "plan_name": update.plan_name,
-            "new_price": update.price,
-            "new_stripe_price_id": new_price.id
-        }
-        
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Failed to update price: {str(e)}")
-
 if __name__ == "__main__":
     import uvicorn
-    uvicorn.run(app, host="0.0.0.0", port=8000) 
-# Widget Add-on Endpoints
-
-@app.get("/check-widget-access")
-async def check_widget_access(user_id: str = Depends(get_current_user_id)):
-    with conn.cursor() as cur:
-        cur.execute("SELECT widget_addon FROM users WHERE id = %s", (user_id,))
-        addon = cur.fetchone()[0]
-        if addon:
-            return {"access": True}
-        raise HTTPException(403, "No access - upgrade")
-
-@app.post("/admin/toggle-addon")
-async def toggle_addon(data: dict = Body(...), admin: str = Depends(get_current_admin)):
-    with conn.cursor() as cur:
-        cur.execute("UPDATE users SET widget_addon = %s WHERE id = %s", (data['enabled'], data['user_id']))
-        conn.commit()
-    return {"status": "updated"}
-
-# AI-Enhanced User Profile Endpoints
-@app.get("/users/profile/enhanced")
-async def get_enhanced_user_profile(user_id: int = Depends(get_current_user_id)):
-    """Get enhanced user profile including AI preferences and recent properties"""
-    try:
-        profile = get_user_profile(user_id)
-        recent_properties = get_recent_properties(user_id, limit=5)
-        
-        return {
-            "profile": profile,
-            "recent_properties": recent_properties,
-            "ai_enabled": profile.get('auto_populate_company_info', True) if profile else True
-        }
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Failed to get enhanced profile: {str(e)}")
-
-@app.post("/users/profile/enhanced")
-async def update_enhanced_user_profile(
-    profile_data: dict = Body(...), 
-    user_id: int = Depends(get_current_user_id)
-):
-    """Update user profile for AI-enhanced deed generation"""
-    try:
-        success = update_user_profile(user_id, profile_data)
-        if success:
-            return {"status": "updated", "message": "Profile updated successfully - AI suggestions will improve!"}
-        else:
-            raise HTTPException(status_code=500, detail="Failed to update profile")
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Profile update failed: {str(e)}")
-
-# Property Caching Endpoints
-@app.post("/property/cache")
-async def cache_property(
-    property_data: dict = Body(...),
-    user_id: int = Depends(get_current_user_id)
-):
-    """Cache property data for future AI suggestions"""
-    try:
-        success = cache_property_data(user_id, property_data)
-        if success:
-            return {"status": "cached", "message": "Property data cached for future suggestions"}
-        else:
-            raise HTTPException(status_code=500, detail="Failed to cache property data")
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Property caching failed: {str(e)}")
-
-@app.get("/property/suggestions")
-async def get_property_suggestions(
-    address: str = Query(..., description="Partial address to search for"),
-    user_id: int = Depends(get_current_user_id)
-):
-    """Get property suggestions based on cached data and address search"""
-    try:
-        # Look for cached properties matching the address
-        cached_property = get_cached_property(user_id, address)
-        recent_properties = get_recent_properties(user_id, limit=3)
-        
-        suggestions = []
-        
-        if cached_property:
-            suggestions.append({
-                "type": "cached_exact",
-                "property": cached_property,
-                "confidence": 0.95
-            })
-        
-        # Add recent properties as suggestions
-        for prop in recent_properties:
-            if address.lower() in prop.get('property_address', '').lower():
-                suggestions.append({
-                    "type": "recent_match",
-                    "property": prop,
-                    "confidence": 0.8
-                })
-        
-        return {
-            "suggestions": suggestions,
-            "total": len(suggestions)
-        }
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Property suggestions failed: {str(e)}")
-
-# AI Suggestions Endpoint (Real-time)
-@app.post("/ai/deed-suggestions")
-async def get_deed_suggestions(
-    deed_data: dict = Body(...),
-    user_id: int = Depends(get_current_user_id)
-):
-    """Get real-time AI suggestions for deed form fields"""
-    try:
-        # Get user profile and cached property data
-        profile = get_user_profile(user_id)
-        cached_property = None
-        
-        if deed_data.get('propertySearch'):
-            cached_property = get_cached_property(user_id, deed_data['propertySearch'])
-        
-        user_data = {
-            'profile': profile,
-            'cached_property': cached_property
-        }
-        
-        # Generate suggestions and validation
-        suggestions = suggest_defaults(user_data, deed_data)
-        validation = validate_deed_data(deed_data, deed_data.get('deedType', 'grant_deed'))
-        
-        return {
-            "suggestions": suggestions,
-            "validation": validation,
-            "profile_available": bool(profile),
-            "cached_data_available": bool(cached_property)
-        }
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"AI suggestions failed: {str(e)}")
+    uvicorn.run(app, host="0.0.0.0", port=8000)
